Compute CARP solution path

CARP returns a fast approximation to the Convex Clustering solution path along with visualizations such as dendrograms and cluster paths. Visualizations may be static, interactive, or both.

CARP(X, obs.labels = NULL, var.labels = NULL, X.center = TRUE,
  X.scale = FALSE, rho = 1, phi = 0.001, weight.dist = c("euclidean",
  "maximum", "manhattan", "canberra", "binary", "minkowski"),
  weight.dist.p = 2, weights = NULL, k = NULL, ncores = 1,
  max.iter = 1e+06, burn.in = 50, verbose = 1, alg.type = c("carpviz",
  "carp", "carpl1", "carpvizl1"), t = 1.05, interactive = TRUE,
  static = TRUE, npcs = 4)

Arguments

X
A n by p matrix with rows the observations and columns the variables

obs.labels
a vector of length n containing observations (row) labels

var.labels
a vector of length p containing variable (column) labels

X.center
A logical. Should X be centered?

X.scale
A logical. Should X be scaled?

rho
A positive number for augmented lagrangian. Not advisable to change.

phi
A positive numner used for scaling in RBF kernel

weight.dist
a string indicating the distance metric used to calculate weights

weight.dist.p
The power of the Minkowski distance, if used.

weights
A vector of positive number of length choose(n,2). Determines observation pairs' fusion amount.

k
an integer >= 1. The number of neighbors used to create sparse weights

ncores
an integer >= 1. The number of cores to use.

max.iter
an integer. The maximum number of CARP iterations.

burn.in
an integer. The number of initial iterations at a fixed value of (small) lambda_k

verbose
either 0,1,or 2. Higher numbers indicate more verbosity while computing.

alg.type
Which CARP algorithm to perform. Choices are 'carpviz' and 'carp;

t
a number greater than 1. The size of the multiplicitive regularization parameter update. Typical values are: 1.1, 1.05, 1.01, 1.005

interactive
A logical. Should interactive paths and dendrograms be returned?

static
A logical. Should static paths and dendrograms be returned?

npcs
A integer >= 2. The number of principal components to compute for path visualization.

Value

X the original data matrix

carp.dend if static==TRUE, a dendrogam representation of the clustering solution path

carp.cluster.path.vis the CARP solution path

n.obs the number of observations

p.vars the number of variables

phi A positive numner used for scaling in RBF kernel

k an integer >= 1. The number of neighbors used to create sparse weights

burn.in an integer. The number of initial iterations at a fixed value of (small) lambda_k

alg.type Which CARP algorithm to perform. Choices are 'carpviz' and 'carp;

X.center A logical. Should X be centered?

X.scale A logical. Should X be scaled?

Details

CARP solves the Convex Clustering problem via Algorithmic Regularization Paths. A seqeunce of clustering solutions is returned along with several visualizations.

Examples

library(clustRviz)
data("presidential_speech")
Xdat <- presidential_speech$X[1:10,1:4]
carp.fit <- CARP(
    X=Xdat,
    obs.labels=presidential_speech$labels[1:10])

#> Pre-computing weight-based edge sets

#> Computing CARP Path

#> Post-processing

carp.fit

#> CARP Fit Summary

#> Number of Observations: 10

#> Number of Variables: 4

#> Pre-processing: center

#> Weights: RBF Kernel, phi = 0.001, k = 2

#> Algorithm: CARP-VIZ

#> Visualizations: Static DendStatic PathInteractive Dend/Path

#> Raw Data:

#>                     amount appropri  british     cent
#> Abraham Lincoln   3.433987 2.397895 1.791759 2.564949
#> Andrew Jackson    4.248495 4.663439 2.995732 1.945910
#> Andrew Johnson    4.025352 3.091042 2.833213 3.332205
#> Barack Obama      1.386294 0.000000 0.000000 1.386294
#> Benjamin Harrison 4.060443 4.174387 2.302585 4.304065