Get clustering solution from a CARP object

Returns cluster labels and cluster means at a point along the CARP path, or the entire cluster sequence of cluster labels and means

# S3 method for CARP
Clustering(x, k = NULL, percent = NULL, ...)

Arguments

x
A CARP object returned by CARP

k
An interger between 1 and n.obs. The number of unique clusters

percent
A number between 0 and 1. The percent of regularization at which to cut the path.

...
Unused additional generic arguements

Value

A list with elements

clustering.assignment

In the case where either k or percent is specified, a vector of cluster labels of length n.obs. In the case where neither k nor percent is specified, a matrix of size n.obs by n.obs, each row specifying a unique cluster assignment along the CARP path.

cluster.means

In the case where either k or percent is specified, a matrix of dimension p.vars by length(unique(clustering.assignment)) with each column a cluster mean. In the case where neither k nor percent is specificed, a list of matricies of length n.obs, representing the cluster means for each cluster assignment along the CARP path.

Details

Passing either the desired number of clusters (k) or the percent regularization (percent) returns the clustering assignment and cluster means at the specific point along the CARP path. If neither k nor percent are specified, all clusteirng assignments and mean matricies are returned.

Examples

library(clustRviz)
data("presidential_speech")
Xdat <- presidential_speech$X[1:10,1:4]
carp.fit <- CARP(
    X=Xdat,
    obs.labels=presidential_speech$labels[1:10])

#> Pre-computing weight-based edge sets

#> Computing CARP Path

#> Post-processing

# Return the CARP iterate with k=5 clusters
carp.clustering <- Clustering(carp.fit,k=5)
# Examine the cluster labels
carp.clustering$clustering.assignment

#>  [1] "cl1" "cl2" "cl2" "cl3" "cl2" "cl2" "cl2" "cl3" "cl4" "cl5"

# Examine the cluster means
head(carp.clustering$cluster.means)

#>           cl1      cl2       cl3       cl4      cl5
#> [1,] 3.433987 3.999821 1.0397208 2.9957323 2.197225
#> [2,] 2.397895 3.970371 0.0000000 3.6635616 2.708050
#> [3,] 1.791759 2.416555 0.0000000 0.6931472 2.944439
#> [4,] 2.564949 3.421564 0.6931472 1.9459101 0.000000

# Return the whole sequence of solutions
carp.clustering.full <- Clustering(carp.fit)
# Examine the k=5 solution again
carp.clustering.full$clustering.assignment[5,]

#>  [1] "cl1" "cl2" "cl2" "cl3" "cl2" "cl2" "cl2" "cl3" "cl4" "cl5"

# Examine the k=5 means again
head(carp.clustering.full$cluster.means[[5]])

#>           cl1      cl2       cl3       cl4      cl5
#> [1,] 3.433987 3.999821 1.0397208 2.9957323 2.197225
#> [2,] 2.397895 3.970371 0.0000000 3.6635616 2.708050
#> [3,] 1.791759 2.416555 0.0000000 0.6931472 2.944439
#> [4,] 2.564949 3.421564 0.6931472 1.9459101 0.000000