Get Clustering Results for CARP

get_cluster_labels returns a factor vector of cluster labels. get_clustered_data returns a matrix (with the same dimensions and names as the original data), but with the values for each observation replaced by its "estimated" value (i.e., the appropriate cluster centroid). get_cluster_centroids returns a k-by-p matrix of cluster centroids, with the same column names as the original data.

get_cluster_labels(x, ...)

# S3 method for CARP
get_cluster_labels(x, ..., percent, k)

get_cluster_centroids(x, ...)

# S3 method for CARP
get_cluster_centroids(x, ..., percent, k, refit = TRUE)

get_clustered_data(x, ...)

# S3 method for CARP
get_clustered_data(x, ..., percent, k, refit = TRUE)

Arguments

x
An object of class CARP as produced by CARP

...
Additional arguments - if any are provided, an error is signalled.

percent
A number between 0 and 1, giving the regularization level (as a fraction of the final regularization level used) at which to get cluster labels.

k
The desired number of clusters. If no iteration with exactly this many clusters is found, the first iterate with fewer than k clusters is used.

refit
Should "naive" centroids (TRUE) or the actual centroids estimated by convex clustering be used? The default (refit = TRUE) centroids returned are actual centroids (mean) of all elements assigned to that cluster; if refit = FALSE, the \(\hat{U}\) from the convex clustering problem is used. Due to the global shrinkage imposed, these clusters are more "shrunk together" than the naive clusters.

Details

get_clustered_data and get_cluster_centroids return centroids on the original scale of the data, independent of any pre-processing flags passed to CARP. Note that exactly one of percent and k must be supplied to each function.

Examples

carp_fit <- CARP(presidential_speech)

#> Pre-computing weights and edge sets

#> Computing Convex Clustering [CARP] Path

#> Post-processing

# Get clustering results from 50% along the path
get_cluster_labels(carp_fit, percent = 0.5)

#>        Abraham Lincoln         Andrew Jackson         Andrew Johnson 
#>              cluster_1              cluster_1              cluster_1 
#>           Barack Obama      Benjamin Harrison        Calvin Coolidge 
#>              cluster_2              cluster_1              cluster_1 
#>      Chester A. Arthur        Donald J. Trump   Dwight D. Eisenhower 
#>              cluster_1              cluster_2              cluster_2 
#>  Franklin D. Roosevelt        Franklin Pierce            George Bush 
#>              cluster_2              cluster_1              cluster_2 
#>         George W. Bush      George Washington         Gerald R. Ford 
#>              cluster_2              cluster_1              cluster_2 
#>       Grover Cleveland        Harry S. Truman         Herbert Hoover 
#>              cluster_1              cluster_2              cluster_1 
#>      James A. Garfield         James Buchanan          James K. Polk 
#>              cluster_1              cluster_1              cluster_1 
#>          James Madison           James Monroe           Jimmy Carter 
#>              cluster_1              cluster_1              cluster_2 
#>             John Adams        John F. Kennedy      John Quincy Adams 
#>              cluster_1              cluster_2              cluster_1 
#>             John Tyler      Lyndon B. Johnson       Martin van Buren 
#>              cluster_1              cluster_2              cluster_1 
#>       Millard Fillmore          Richard Nixon          Ronald Reagan 
#>              cluster_1              cluster_2              cluster_2 
#>    Rutherford B. Hayes     Theodore Roosevelt       Thomas Jefferson 
#>              cluster_1              cluster_1              cluster_1 
#>       Ulysses S. Grant      Warren G. Harding William Henry Harrison 
#>              cluster_1              cluster_1              cluster_1 
#>    William Howard Taft     William J. Clinton       William McKinley 
#>              cluster_1              cluster_2              cluster_1 
#>         Woodrow Wilson         Zachary Taylor 
#>              cluster_1              cluster_1 
#> Levels: cluster_1 cluster_2

# Get labels corresponding to the 3 cluster solution
get_cluster_labels(carp_fit, k = 3)

#>        Abraham Lincoln         Andrew Jackson         Andrew Johnson 
#>              cluster_1              cluster_1              cluster_1 
#>           Barack Obama      Benjamin Harrison        Calvin Coolidge 
#>              cluster_2              cluster_1              cluster_1 
#>      Chester A. Arthur        Donald J. Trump   Dwight D. Eisenhower 
#>              cluster_1              cluster_2              cluster_2 
#>  Franklin D. Roosevelt        Franklin Pierce            George Bush 
#>              cluster_2              cluster_1              cluster_2 
#>         George W. Bush      George Washington         Gerald R. Ford 
#>              cluster_2              cluster_1              cluster_2 
#>       Grover Cleveland        Harry S. Truman         Herbert Hoover 
#>              cluster_1              cluster_2              cluster_1 
#>      James A. Garfield         James Buchanan          James K. Polk 
#>              cluster_1              cluster_1              cluster_1 
#>          James Madison           James Monroe           Jimmy Carter 
#>              cluster_1              cluster_1              cluster_2 
#>             John Adams        John F. Kennedy      John Quincy Adams 
#>              cluster_1              cluster_2              cluster_1 
#>             John Tyler      Lyndon B. Johnson       Martin van Buren 
#>              cluster_1              cluster_2              cluster_1 
#>       Millard Fillmore          Richard Nixon          Ronald Reagan 
#>              cluster_1              cluster_2              cluster_2 
#>    Rutherford B. Hayes     Theodore Roosevelt       Thomas Jefferson 
#>              cluster_1              cluster_1              cluster_1 
#>       Ulysses S. Grant      Warren G. Harding William Henry Harrison 
#>              cluster_1              cluster_3              cluster_1 
#>    William Howard Taft     William J. Clinton       William McKinley 
#>              cluster_1              cluster_2              cluster_1 
#>         Woodrow Wilson         Zachary Taylor 
#>              cluster_1              cluster_1 
#> Levels: cluster_1 cluster_2 cluster_3

# Get 3 cluster centroids
get_cluster_centroids(carp_fit, k = 3)

#>        amount  appropri   british      cent  commerci commission consider
#> [1,] 3.359010 3.3140457 2.3921950 2.0502205 2.7591423  2.4623515 3.667895
#> [2,] 1.862722 2.0425014 0.5157932 0.7925161 0.9780746  0.1980421 1.465250
#> [3,] 0.000000 0.6931472 0.0000000 2.1972246 2.3025851  0.0000000 2.708050
#>       develop expenditur     farm    feder    fiscal      help   indian
#> [1,] 2.332016   3.127940 1.160957 2.357668 2.4904667 0.8689617 2.868423
#> [2,] 3.715357   1.996894 2.392863 4.032946 2.5822486 4.3600998 1.028310
#> [3,] 2.484907   1.945910 2.079442 2.995732 0.6931472 2.7080502 0.000000
#>      intercours    island     june      mail    mexico      navi     need
#> [1,]   2.427929 2.3262376 2.283998 2.1068540 2.4916409 2.9538237 2.602382
#> [2,]   0.000000 0.9483649 1.273441 0.4409847 0.7222947 0.9645195 4.444242
#> [3,]   0.000000 0.0000000 0.000000 0.0000000 0.0000000 0.0000000 2.833213
#>           per   provis   receipt   reform   regard   report    shall  subject
#> [1,] 2.038704 3.342505 2.4833353 1.589620 2.980658 3.231835 3.893628 3.969144
#> [2,] 1.529186 1.351412 0.7369229 2.950234 1.399618 2.504706 2.266559 1.660877
#> [3,] 2.484907 2.564949 0.0000000 1.098612 0.000000 1.945910 3.295837 1.098612
#>      territori treasuri   treati     upon    vessel  america     bank     incom
#> [1,]  3.262769 3.321236 3.527688 4.526728 2.9735118 2.087653 2.511195 1.3637360
#> [2,]  1.031974 1.170382 1.850235 2.707913 0.5064341 4.545829 1.961930 3.3394351
#> [3,]  1.609438 1.791759 1.945910 2.833213 0.0000000 2.890372 1.098612 0.6931472
#>       million    price     spain    articl       cut    dollar   econom
#> [1,] 1.794344 2.086668 2.0103993 2.4389560 0.7648271 1.5076055 1.609623
#> [2,] 4.044070 3.083916 0.1980421 0.2475526 3.1468759 3.2737465 3.999372
#> [3,] 1.791759 1.945910 0.0000000 0.6931472 1.0986123 0.6931472 2.708050
#>          women     educ   school      bill    tariff  unemploy    weapon
#> [1,] 0.5746917 1.892278 1.291019 1.9113003 1.9128817 0.3871545 0.2768995
#> [2,] 2.9420389 3.463114 3.235803 2.8435069 0.8874667 2.7956399 2.7877462
#> [3,] 1.0986123 2.197225 1.945910 0.6931472 2.8332133 1.6094379 0.0000000
#>            get   method   farmer  challeng   achiev democraci      area
#> [1,] 0.5113785 1.514215 1.093850 0.3427457 1.049185 0.2999833 0.9618479
#> [2,] 3.0255971 1.461203 2.349569 3.1210443 3.403615 2.5199520 3.0214103
#> [3,] 0.6931472 1.386294 2.302585 1.6094379 1.386294 1.6094379 1.0986123
#>        billion  problem     basic   budget      goal       job     level
#> [1,] 0.2441388 1.162271 0.1109957 0.430126 0.1033011 0.1474712 0.5037702
#> [2,] 3.7429589 3.659555 2.8387802 3.795996 3.2118895 3.9587172 3.0683132
#> [3,] 1.3862944 3.526361 1.0986123 1.386294 0.6931472 0.0000000 1.7917595
#>       nuclear   percent   program     spend  technolog     today  tonight
#> [1,] 0.000000 0.1527868 0.4188332 0.2746740 0.02390163 0.2544055 0.000000
#> [2,] 2.559875 3.4095268 4.4328005 3.1180550 2.68228351 3.7233467 2.596381
#> [3,] 0.000000 0.0000000 2.1972246 0.6931472 0.00000000 2.3978953 0.000000
#>         worker    inflat   soviet
#> [1,] 0.2961965 0.3494689 0.000000
#> [2,] 3.1347856 2.5319042 2.493221
#> [3,] 0.6931472 1.6094379 0.000000

# Get the clustered estimates for k = 3 clusters
get_clustered_data(carp_fit, k = 3)

#>                          amount  appropri   british      cent  commerci
#> Abraham Lincoln        3.359010 3.3140457 2.3921950 2.0502205 2.7591423
#> Andrew Jackson         3.359010 3.3140457 2.3921950 2.0502205 2.7591423
#> Andrew Johnson         3.359010 3.3140457 2.3921950 2.0502205 2.7591423
#> Barack Obama           1.862722 2.0425014 0.5157932 0.7925161 0.9780746
#> Benjamin Harrison      3.359010 3.3140457 2.3921950 2.0502205 2.7591423
#> Calvin Coolidge        3.359010 3.3140457 2.3921950 2.0502205 2.7591423
#> Chester A. Arthur      3.359010 3.3140457 2.3921950 2.0502205 2.7591423
#> Donald J. Trump        1.862722 2.0425014 0.5157932 0.7925161 0.9780746
#> Dwight D. Eisenhower   1.862722 2.0425014 0.5157932 0.7925161 0.9780746
#> Franklin D. Roosevelt  1.862722 2.0425014 0.5157932 0.7925161 0.9780746
#> Franklin Pierce        3.359010 3.3140457 2.3921950 2.0502205 2.7591423
#> George Bush            1.862722 2.0425014 0.5157932 0.7925161 0.9780746
#> George W. Bush         1.862722 2.0425014 0.5157932 0.7925161 0.9780746
#> George Washington      3.359010 3.3140457 2.3921950 2.0502205 2.7591423
#> Gerald R. Ford         1.862722 2.0425014 0.5157932 0.7925161 0.9780746
#> Grover Cleveland       3.359010 3.3140457 2.3921950 2.0502205 2.7591423
#> Harry S. Truman        1.862722 2.0425014 0.5157932 0.7925161 0.9780746
#> Herbert Hoover         3.359010 3.3140457 2.3921950 2.0502205 2.7591423
#> James A. Garfield      3.359010 3.3140457 2.3921950 2.0502205 2.7591423
#> James Buchanan         3.359010 3.3140457 2.3921950 2.0502205 2.7591423
#> James K. Polk          3.359010 3.3140457 2.3921950 2.0502205 2.7591423
#> James Madison          3.359010 3.3140457 2.3921950 2.0502205 2.7591423
#> James Monroe           3.359010 3.3140457 2.3921950 2.0502205 2.7591423
#> Jimmy Carter           1.862722 2.0425014 0.5157932 0.7925161 0.9780746
#> John Adams             3.359010 3.3140457 2.3921950 2.0502205 2.7591423
#> John F. Kennedy        1.862722 2.0425014 0.5157932 0.7925161 0.9780746
#> John Quincy Adams      3.359010 3.3140457 2.3921950 2.0502205 2.7591423
#> John Tyler             3.359010 3.3140457 2.3921950 2.0502205 2.7591423
#> Lyndon B. Johnson      1.862722 2.0425014 0.5157932 0.7925161 0.9780746
#> Martin van Buren       3.359010 3.3140457 2.3921950 2.0502205 2.7591423
#> Millard Fillmore       3.359010 3.3140457 2.3921950 2.0502205 2.7591423
#> Richard Nixon          1.862722 2.0425014 0.5157932 0.7925161 0.9780746
#> Ronald Reagan          1.862722 2.0425014 0.5157932 0.7925161 0.9780746
#> Rutherford B. Hayes    3.359010 3.3140457 2.3921950 2.0502205 2.7591423
#> Theodore Roosevelt     3.359010 3.3140457 2.3921950 2.0502205 2.7591423
#> Thomas Jefferson       3.359010 3.3140457 2.3921950 2.0502205 2.7591423
#> Ulysses S. Grant       3.359010 3.3140457 2.3921950 2.0502205 2.7591423
#> Warren G. Harding      0.000000 0.6931472 0.0000000 2.1972246 2.3025851
#> William Henry Harrison 3.359010 3.3140457 2.3921950 2.0502205 2.7591423
#> William Howard Taft    3.359010 3.3140457 2.3921950 2.0502205 2.7591423
#> William J. Clinton     1.862722 2.0425014 0.5157932 0.7925161 0.9780746
#> William McKinley       3.359010 3.3140457 2.3921950 2.0502205 2.7591423
#> Woodrow Wilson         3.359010 3.3140457 2.3921950 2.0502205 2.7591423
#> Zachary Taylor         3.359010 3.3140457 2.3921950 2.0502205 2.7591423
#>                        commission consider  develop expenditur     farm
#> Abraham Lincoln         2.4623515 3.667895 2.332016   3.127940 1.160957
#> Andrew Jackson          2.4623515 3.667895 2.332016   3.127940 1.160957
#> Andrew Johnson          2.4623515 3.667895 2.332016   3.127940 1.160957
#> Barack Obama            0.1980421 1.465250 3.715357   1.996894 2.392863
#> Benjamin Harrison       2.4623515 3.667895 2.332016   3.127940 1.160957
#> Calvin Coolidge         2.4623515 3.667895 2.332016   3.127940 1.160957
#> Chester A. Arthur       2.4623515 3.667895 2.332016   3.127940 1.160957
#> Donald J. Trump         0.1980421 1.465250 3.715357   1.996894 2.392863
#> Dwight D. Eisenhower    0.1980421 1.465250 3.715357   1.996894 2.392863
#> Franklin D. Roosevelt   0.1980421 1.465250 3.715357   1.996894 2.392863
#> Franklin Pierce         2.4623515 3.667895 2.332016   3.127940 1.160957
#> George Bush             0.1980421 1.465250 3.715357   1.996894 2.392863
#> George W. Bush          0.1980421 1.465250 3.715357   1.996894 2.392863
#> George Washington       2.4623515 3.667895 2.332016   3.127940 1.160957
#> Gerald R. Ford          0.1980421 1.465250 3.715357   1.996894 2.392863
#> Grover Cleveland        2.4623515 3.667895 2.332016   3.127940 1.160957
#> Harry S. Truman         0.1980421 1.465250 3.715357   1.996894 2.392863
#> Herbert Hoover          2.4623515 3.667895 2.332016   3.127940 1.160957
#> James A. Garfield       2.4623515 3.667895 2.332016   3.127940 1.160957
#> James Buchanan          2.4623515 3.667895 2.332016   3.127940 1.160957
#> James K. Polk           2.4623515 3.667895 2.332016   3.127940 1.160957
#> James Madison           2.4623515 3.667895 2.332016   3.127940 1.160957
#> James Monroe            2.4623515 3.667895 2.332016   3.127940 1.160957
#> Jimmy Carter            0.1980421 1.465250 3.715357   1.996894 2.392863
#> John Adams              2.4623515 3.667895 2.332016   3.127940 1.160957
#> John F. Kennedy         0.1980421 1.465250 3.715357   1.996894 2.392863
#> John Quincy Adams       2.4623515 3.667895 2.332016   3.127940 1.160957
#> John Tyler              2.4623515 3.667895 2.332016   3.127940 1.160957
#> Lyndon B. Johnson       0.1980421 1.465250 3.715357   1.996894 2.392863
#> Martin van Buren        2.4623515 3.667895 2.332016   3.127940 1.160957
#> Millard Fillmore        2.4623515 3.667895 2.332016   3.127940 1.160957
#> Richard Nixon           0.1980421 1.465250 3.715357   1.996894 2.392863
#> Ronald Reagan           0.1980421 1.465250 3.715357   1.996894 2.392863
#> Rutherford B. Hayes     2.4623515 3.667895 2.332016   3.127940 1.160957
#> Theodore Roosevelt      2.4623515 3.667895 2.332016   3.127940 1.160957
#> Thomas Jefferson        2.4623515 3.667895 2.332016   3.127940 1.160957
#> Ulysses S. Grant        2.4623515 3.667895 2.332016   3.127940 1.160957
#> Warren G. Harding       0.0000000 2.708050 2.484907   1.945910 2.079442
#> William Henry Harrison  2.4623515 3.667895 2.332016   3.127940 1.160957
#> William Howard Taft     2.4623515 3.667895 2.332016   3.127940 1.160957
#> William J. Clinton      0.1980421 1.465250 3.715357   1.996894 2.392863
#> William McKinley        2.4623515 3.667895 2.332016   3.127940 1.160957
#> Woodrow Wilson          2.4623515 3.667895 2.332016   3.127940 1.160957
#> Zachary Taylor          2.4623515 3.667895 2.332016   3.127940 1.160957
#>                           feder    fiscal      help   indian intercours
#> Abraham Lincoln        2.357668 2.4904667 0.8689617 2.868423   2.427929
#> Andrew Jackson         2.357668 2.4904667 0.8689617 2.868423   2.427929
#> Andrew Johnson         2.357668 2.4904667 0.8689617 2.868423   2.427929
#> Barack Obama           4.032946 2.5822486 4.3600998 1.028310   0.000000
#> Benjamin Harrison      2.357668 2.4904667 0.8689617 2.868423   2.427929
#> Calvin Coolidge        2.357668 2.4904667 0.8689617 2.868423   2.427929
#> Chester A. Arthur      2.357668 2.4904667 0.8689617 2.868423   2.427929
#> Donald J. Trump        4.032946 2.5822486 4.3600998 1.028310   0.000000
#> Dwight D. Eisenhower   4.032946 2.5822486 4.3600998 1.028310   0.000000
#> Franklin D. Roosevelt  4.032946 2.5822486 4.3600998 1.028310   0.000000
#> Franklin Pierce        2.357668 2.4904667 0.8689617 2.868423   2.427929
#> George Bush            4.032946 2.5822486 4.3600998 1.028310   0.000000
#> George W. Bush         4.032946 2.5822486 4.3600998 1.028310   0.000000
#> George Washington      2.357668 2.4904667 0.8689617 2.868423   2.427929
#> Gerald R. Ford         4.032946 2.5822486 4.3600998 1.028310   0.000000
#> Grover Cleveland       2.357668 2.4904667 0.8689617 2.868423   2.427929
#> Harry S. Truman        4.032946 2.5822486 4.3600998 1.028310   0.000000
#> Herbert Hoover         2.357668 2.4904667 0.8689617 2.868423   2.427929
#> James A. Garfield      2.357668 2.4904667 0.8689617 2.868423   2.427929
#> James Buchanan         2.357668 2.4904667 0.8689617 2.868423   2.427929
#> James K. Polk          2.357668 2.4904667 0.8689617 2.868423   2.427929
#> James Madison          2.357668 2.4904667 0.8689617 2.868423   2.427929
#> James Monroe           2.357668 2.4904667 0.8689617 2.868423   2.427929
#> Jimmy Carter           4.032946 2.5822486 4.3600998 1.028310   0.000000
#> John Adams             2.357668 2.4904667 0.8689617 2.868423   2.427929
#> John F. Kennedy        4.032946 2.5822486 4.3600998 1.028310   0.000000
#> John Quincy Adams      2.357668 2.4904667 0.8689617 2.868423   2.427929
#> John Tyler             2.357668 2.4904667 0.8689617 2.868423   2.427929
#> Lyndon B. Johnson      4.032946 2.5822486 4.3600998 1.028310   0.000000
#> Martin van Buren       2.357668 2.4904667 0.8689617 2.868423   2.427929
#> Millard Fillmore       2.357668 2.4904667 0.8689617 2.868423   2.427929
#> Richard Nixon          4.032946 2.5822486 4.3600998 1.028310   0.000000
#> Ronald Reagan          4.032946 2.5822486 4.3600998 1.028310   0.000000
#> Rutherford B. Hayes    2.357668 2.4904667 0.8689617 2.868423   2.427929
#> Theodore Roosevelt     2.357668 2.4904667 0.8689617 2.868423   2.427929
#> Thomas Jefferson       2.357668 2.4904667 0.8689617 2.868423   2.427929
#> Ulysses S. Grant       2.357668 2.4904667 0.8689617 2.868423   2.427929
#> Warren G. Harding      2.995732 0.6931472 2.7080502 0.000000   0.000000
#> William Henry Harrison 2.357668 2.4904667 0.8689617 2.868423   2.427929
#> William Howard Taft    2.357668 2.4904667 0.8689617 2.868423   2.427929
#> William J. Clinton     4.032946 2.5822486 4.3600998 1.028310   0.000000
#> William McKinley       2.357668 2.4904667 0.8689617 2.868423   2.427929
#> Woodrow Wilson         2.357668 2.4904667 0.8689617 2.868423   2.427929
#> Zachary Taylor         2.357668 2.4904667 0.8689617 2.868423   2.427929
#>                           island     june      mail    mexico      navi
#> Abraham Lincoln        2.3262376 2.283998 2.1068540 2.4916409 2.9538237
#> Andrew Jackson         2.3262376 2.283998 2.1068540 2.4916409 2.9538237
#> Andrew Johnson         2.3262376 2.283998 2.1068540 2.4916409 2.9538237
#> Barack Obama           0.9483649 1.273441 0.4409847 0.7222947 0.9645195
#> Benjamin Harrison      2.3262376 2.283998 2.1068540 2.4916409 2.9538237
#> Calvin Coolidge        2.3262376 2.283998 2.1068540 2.4916409 2.9538237
#> Chester A. Arthur      2.3262376 2.283998 2.1068540 2.4916409 2.9538237
#> Donald J. Trump        0.9483649 1.273441 0.4409847 0.7222947 0.9645195
#> Dwight D. Eisenhower   0.9483649 1.273441 0.4409847 0.7222947 0.9645195
#> Franklin D. Roosevelt  0.9483649 1.273441 0.4409847 0.7222947 0.9645195
#> Franklin Pierce        2.3262376 2.283998 2.1068540 2.4916409 2.9538237
#> George Bush            0.9483649 1.273441 0.4409847 0.7222947 0.9645195
#> George W. Bush         0.9483649 1.273441 0.4409847 0.7222947 0.9645195
#> George Washington      2.3262376 2.283998 2.1068540 2.4916409 2.9538237
#> Gerald R. Ford         0.9483649 1.273441 0.4409847 0.7222947 0.9645195
#> Grover Cleveland       2.3262376 2.283998 2.1068540 2.4916409 2.9538237
#> Harry S. Truman        0.9483649 1.273441 0.4409847 0.7222947 0.9645195
#> Herbert Hoover         2.3262376 2.283998 2.1068540 2.4916409 2.9538237
#> James A. Garfield      2.3262376 2.283998 2.1068540 2.4916409 2.9538237
#> James Buchanan         2.3262376 2.283998 2.1068540 2.4916409 2.9538237
#> James K. Polk          2.3262376 2.283998 2.1068540 2.4916409 2.9538237
#> James Madison          2.3262376 2.283998 2.1068540 2.4916409 2.9538237
#> James Monroe           2.3262376 2.283998 2.1068540 2.4916409 2.9538237
#> Jimmy Carter           0.9483649 1.273441 0.4409847 0.7222947 0.9645195
#> John Adams             2.3262376 2.283998 2.1068540 2.4916409 2.9538237
#> John F. Kennedy        0.9483649 1.273441 0.4409847 0.7222947 0.9645195
#> John Quincy Adams      2.3262376 2.283998 2.1068540 2.4916409 2.9538237
#> John Tyler             2.3262376 2.283998 2.1068540 2.4916409 2.9538237
#> Lyndon B. Johnson      0.9483649 1.273441 0.4409847 0.7222947 0.9645195
#> Martin van Buren       2.3262376 2.283998 2.1068540 2.4916409 2.9538237
#> Millard Fillmore       2.3262376 2.283998 2.1068540 2.4916409 2.9538237
#> Richard Nixon          0.9483649 1.273441 0.4409847 0.7222947 0.9645195
#> Ronald Reagan          0.9483649 1.273441 0.4409847 0.7222947 0.9645195
#> Rutherford B. Hayes    2.3262376 2.283998 2.1068540 2.4916409 2.9538237
#> Theodore Roosevelt     2.3262376 2.283998 2.1068540 2.4916409 2.9538237
#> Thomas Jefferson       2.3262376 2.283998 2.1068540 2.4916409 2.9538237
#> Ulysses S. Grant       2.3262376 2.283998 2.1068540 2.4916409 2.9538237
#> Warren G. Harding      0.0000000 0.000000 0.0000000 0.0000000 0.0000000
#> William Henry Harrison 2.3262376 2.283998 2.1068540 2.4916409 2.9538237
#> William Howard Taft    2.3262376 2.283998 2.1068540 2.4916409 2.9538237
#> William J. Clinton     0.9483649 1.273441 0.4409847 0.7222947 0.9645195
#> William McKinley       2.3262376 2.283998 2.1068540 2.4916409 2.9538237
#> Woodrow Wilson         2.3262376 2.283998 2.1068540 2.4916409 2.9538237
#> Zachary Taylor         2.3262376 2.283998 2.1068540 2.4916409 2.9538237
#>                            need      per   provis   receipt   reform   regard
#> Abraham Lincoln        2.602382 2.038704 3.342505 2.4833353 1.589620 2.980658
#> Andrew Jackson         2.602382 2.038704 3.342505 2.4833353 1.589620 2.980658
#> Andrew Johnson         2.602382 2.038704 3.342505 2.4833353 1.589620 2.980658
#> Barack Obama           4.444242 1.529186 1.351412 0.7369229 2.950234 1.399618
#> Benjamin Harrison      2.602382 2.038704 3.342505 2.4833353 1.589620 2.980658
#> Calvin Coolidge        2.602382 2.038704 3.342505 2.4833353 1.589620 2.980658
#> Chester A. Arthur      2.602382 2.038704 3.342505 2.4833353 1.589620 2.980658
#> Donald J. Trump        4.444242 1.529186 1.351412 0.7369229 2.950234 1.399618
#> Dwight D. Eisenhower   4.444242 1.529186 1.351412 0.7369229 2.950234 1.399618
#> Franklin D. Roosevelt  4.444242 1.529186 1.351412 0.7369229 2.950234 1.399618
#> Franklin Pierce        2.602382 2.038704 3.342505 2.4833353 1.589620 2.980658
#> George Bush            4.444242 1.529186 1.351412 0.7369229 2.950234 1.399618
#> George W. Bush         4.444242 1.529186 1.351412 0.7369229 2.950234 1.399618
#> George Washington      2.602382 2.038704 3.342505 2.4833353 1.589620 2.980658
#> Gerald R. Ford         4.444242 1.529186 1.351412 0.7369229 2.950234 1.399618
#> Grover Cleveland       2.602382 2.038704 3.342505 2.4833353 1.589620 2.980658
#> Harry S. Truman        4.444242 1.529186 1.351412 0.7369229 2.950234 1.399618
#> Herbert Hoover         2.602382 2.038704 3.342505 2.4833353 1.589620 2.980658
#> James A. Garfield      2.602382 2.038704 3.342505 2.4833353 1.589620 2.980658
#> James Buchanan         2.602382 2.038704 3.342505 2.4833353 1.589620 2.980658
#> James K. Polk          2.602382 2.038704 3.342505 2.4833353 1.589620 2.980658
#> James Madison          2.602382 2.038704 3.342505 2.4833353 1.589620 2.980658
#> James Monroe           2.602382 2.038704 3.342505 2.4833353 1.589620 2.980658
#> Jimmy Carter           4.444242 1.529186 1.351412 0.7369229 2.950234 1.399618
#> John Adams             2.602382 2.038704 3.342505 2.4833353 1.589620 2.980658
#> John F. Kennedy        4.444242 1.529186 1.351412 0.7369229 2.950234 1.399618
#> John Quincy Adams      2.602382 2.038704 3.342505 2.4833353 1.589620 2.980658
#> John Tyler             2.602382 2.038704 3.342505 2.4833353 1.589620 2.980658
#> Lyndon B. Johnson      4.444242 1.529186 1.351412 0.7369229 2.950234 1.399618
#> Martin van Buren       2.602382 2.038704 3.342505 2.4833353 1.589620 2.980658
#> Millard Fillmore       2.602382 2.038704 3.342505 2.4833353 1.589620 2.980658
#> Richard Nixon          4.444242 1.529186 1.351412 0.7369229 2.950234 1.399618
#> Ronald Reagan          4.444242 1.529186 1.351412 0.7369229 2.950234 1.399618
#> Rutherford B. Hayes    2.602382 2.038704 3.342505 2.4833353 1.589620 2.980658
#> Theodore Roosevelt     2.602382 2.038704 3.342505 2.4833353 1.589620 2.980658
#> Thomas Jefferson       2.602382 2.038704 3.342505 2.4833353 1.589620 2.980658
#> Ulysses S. Grant       2.602382 2.038704 3.342505 2.4833353 1.589620 2.980658
#> Warren G. Harding      2.833213 2.484907 2.564949 0.0000000 1.098612 0.000000
#> William Henry Harrison 2.602382 2.038704 3.342505 2.4833353 1.589620 2.980658
#> William Howard Taft    2.602382 2.038704 3.342505 2.4833353 1.589620 2.980658
#> William J. Clinton     4.444242 1.529186 1.351412 0.7369229 2.950234 1.399618
#> William McKinley       2.602382 2.038704 3.342505 2.4833353 1.589620 2.980658
#> Woodrow Wilson         2.602382 2.038704 3.342505 2.4833353 1.589620 2.980658
#> Zachary Taylor         2.602382 2.038704 3.342505 2.4833353 1.589620 2.980658
#>                          report    shall  subject territori treasuri   treati
#> Abraham Lincoln        3.231835 3.893628 3.969144  3.262769 3.321236 3.527688
#> Andrew Jackson         3.231835 3.893628 3.969144  3.262769 3.321236 3.527688
#> Andrew Johnson         3.231835 3.893628 3.969144  3.262769 3.321236 3.527688
#> Barack Obama           2.504706 2.266559 1.660877  1.031974 1.170382 1.850235
#> Benjamin Harrison      3.231835 3.893628 3.969144  3.262769 3.321236 3.527688
#> Calvin Coolidge        3.231835 3.893628 3.969144  3.262769 3.321236 3.527688
#> Chester A. Arthur      3.231835 3.893628 3.969144  3.262769 3.321236 3.527688
#> Donald J. Trump        2.504706 2.266559 1.660877  1.031974 1.170382 1.850235
#> Dwight D. Eisenhower   2.504706 2.266559 1.660877  1.031974 1.170382 1.850235
#> Franklin D. Roosevelt  2.504706 2.266559 1.660877  1.031974 1.170382 1.850235
#> Franklin Pierce        3.231835 3.893628 3.969144  3.262769 3.321236 3.527688
#> George Bush            2.504706 2.266559 1.660877  1.031974 1.170382 1.850235
#> George W. Bush         2.504706 2.266559 1.660877  1.031974 1.170382 1.850235
#> George Washington      3.231835 3.893628 3.969144  3.262769 3.321236 3.527688
#> Gerald R. Ford         2.504706 2.266559 1.660877  1.031974 1.170382 1.850235
#> Grover Cleveland       3.231835 3.893628 3.969144  3.262769 3.321236 3.527688
#> Harry S. Truman        2.504706 2.266559 1.660877  1.031974 1.170382 1.850235
#> Herbert Hoover         3.231835 3.893628 3.969144  3.262769 3.321236 3.527688
#> James A. Garfield      3.231835 3.893628 3.969144  3.262769 3.321236 3.527688
#> James Buchanan         3.231835 3.893628 3.969144  3.262769 3.321236 3.527688
#> James K. Polk          3.231835 3.893628 3.969144  3.262769 3.321236 3.527688
#> James Madison          3.231835 3.893628 3.969144  3.262769 3.321236 3.527688
#> James Monroe           3.231835 3.893628 3.969144  3.262769 3.321236 3.527688
#> Jimmy Carter           2.504706 2.266559 1.660877  1.031974 1.170382 1.850235
#> John Adams             3.231835 3.893628 3.969144  3.262769 3.321236 3.527688
#> John F. Kennedy        2.504706 2.266559 1.660877  1.031974 1.170382 1.850235
#> John Quincy Adams      3.231835 3.893628 3.969144  3.262769 3.321236 3.527688
#> John Tyler             3.231835 3.893628 3.969144  3.262769 3.321236 3.527688
#> Lyndon B. Johnson      2.504706 2.266559 1.660877  1.031974 1.170382 1.850235
#> Martin van Buren       3.231835 3.893628 3.969144  3.262769 3.321236 3.527688
#> Millard Fillmore       3.231835 3.893628 3.969144  3.262769 3.321236 3.527688
#> Richard Nixon          2.504706 2.266559 1.660877  1.031974 1.170382 1.850235
#> Ronald Reagan          2.504706 2.266559 1.660877  1.031974 1.170382 1.850235
#> Rutherford B. Hayes    3.231835 3.893628 3.969144  3.262769 3.321236 3.527688
#> Theodore Roosevelt     3.231835 3.893628 3.969144  3.262769 3.321236 3.527688
#> Thomas Jefferson       3.231835 3.893628 3.969144  3.262769 3.321236 3.527688
#> Ulysses S. Grant       3.231835 3.893628 3.969144  3.262769 3.321236 3.527688
#> Warren G. Harding      1.945910 3.295837 1.098612  1.609438 1.791759 1.945910
#> William Henry Harrison 3.231835 3.893628 3.969144  3.262769 3.321236 3.527688
#> William Howard Taft    3.231835 3.893628 3.969144  3.262769 3.321236 3.527688
#> William J. Clinton     2.504706 2.266559 1.660877  1.031974 1.170382 1.850235
#> William McKinley       3.231835 3.893628 3.969144  3.262769 3.321236 3.527688
#> Woodrow Wilson         3.231835 3.893628 3.969144  3.262769 3.321236 3.527688
#> Zachary Taylor         3.231835 3.893628 3.969144  3.262769 3.321236 3.527688
#>                            upon    vessel  america     bank     incom  million
#> Abraham Lincoln        4.526728 2.9735118 2.087653 2.511195 1.3637360 1.794344
#> Andrew Jackson         4.526728 2.9735118 2.087653 2.511195 1.3637360 1.794344
#> Andrew Johnson         4.526728 2.9735118 2.087653 2.511195 1.3637360 1.794344
#> Barack Obama           2.707913 0.5064341 4.545829 1.961930 3.3394351 4.044070
#> Benjamin Harrison      4.526728 2.9735118 2.087653 2.511195 1.3637360 1.794344
#> Calvin Coolidge        4.526728 2.9735118 2.087653 2.511195 1.3637360 1.794344
#> Chester A. Arthur      4.526728 2.9735118 2.087653 2.511195 1.3637360 1.794344
#> Donald J. Trump        2.707913 0.5064341 4.545829 1.961930 3.3394351 4.044070
#> Dwight D. Eisenhower   2.707913 0.5064341 4.545829 1.961930 3.3394351 4.044070
#> Franklin D. Roosevelt  2.707913 0.5064341 4.545829 1.961930 3.3394351 4.044070
#> Franklin Pierce        4.526728 2.9735118 2.087653 2.511195 1.3637360 1.794344
#> George Bush            2.707913 0.5064341 4.545829 1.961930 3.3394351 4.044070
#> George W. Bush         2.707913 0.5064341 4.545829 1.961930 3.3394351 4.044070
#> George Washington      4.526728 2.9735118 2.087653 2.511195 1.3637360 1.794344
#> Gerald R. Ford         2.707913 0.5064341 4.545829 1.961930 3.3394351 4.044070
#> Grover Cleveland       4.526728 2.9735118 2.087653 2.511195 1.3637360 1.794344
#> Harry S. Truman        2.707913 0.5064341 4.545829 1.961930 3.3394351 4.044070
#> Herbert Hoover         4.526728 2.9735118 2.087653 2.511195 1.3637360 1.794344
#> James A. Garfield      4.526728 2.9735118 2.087653 2.511195 1.3637360 1.794344
#> James Buchanan         4.526728 2.9735118 2.087653 2.511195 1.3637360 1.794344
#> James K. Polk          4.526728 2.9735118 2.087653 2.511195 1.3637360 1.794344
#> James Madison          4.526728 2.9735118 2.087653 2.511195 1.3637360 1.794344
#> James Monroe           4.526728 2.9735118 2.087653 2.511195 1.3637360 1.794344
#> Jimmy Carter           2.707913 0.5064341 4.545829 1.961930 3.3394351 4.044070
#> John Adams             4.526728 2.9735118 2.087653 2.511195 1.3637360 1.794344
#> John F. Kennedy        2.707913 0.5064341 4.545829 1.961930 3.3394351 4.044070
#> John Quincy Adams      4.526728 2.9735118 2.087653 2.511195 1.3637360 1.794344
#> John Tyler             4.526728 2.9735118 2.087653 2.511195 1.3637360 1.794344
#> Lyndon B. Johnson      2.707913 0.5064341 4.545829 1.961930 3.3394351 4.044070
#> Martin van Buren       4.526728 2.9735118 2.087653 2.511195 1.3637360 1.794344
#> Millard Fillmore       4.526728 2.9735118 2.087653 2.511195 1.3637360 1.794344
#> Richard Nixon          2.707913 0.5064341 4.545829 1.961930 3.3394351 4.044070
#> Ronald Reagan          2.707913 0.5064341 4.545829 1.961930 3.3394351 4.044070
#> Rutherford B. Hayes    4.526728 2.9735118 2.087653 2.511195 1.3637360 1.794344
#> Theodore Roosevelt     4.526728 2.9735118 2.087653 2.511195 1.3637360 1.794344
#> Thomas Jefferson       4.526728 2.9735118 2.087653 2.511195 1.3637360 1.794344
#> Ulysses S. Grant       4.526728 2.9735118 2.087653 2.511195 1.3637360 1.794344
#> Warren G. Harding      2.833213 0.0000000 2.890372 1.098612 0.6931472 1.791759
#> William Henry Harrison 4.526728 2.9735118 2.087653 2.511195 1.3637360 1.794344
#> William Howard Taft    4.526728 2.9735118 2.087653 2.511195 1.3637360 1.794344
#> William J. Clinton     2.707913 0.5064341 4.545829 1.961930 3.3394351 4.044070
#> William McKinley       4.526728 2.9735118 2.087653 2.511195 1.3637360 1.794344
#> Woodrow Wilson         4.526728 2.9735118 2.087653 2.511195 1.3637360 1.794344
#> Zachary Taylor         4.526728 2.9735118 2.087653 2.511195 1.3637360 1.794344
#>                           price     spain    articl       cut    dollar
#> Abraham Lincoln        2.086668 2.0103993 2.4389560 0.7648271 1.5076055
#> Andrew Jackson         2.086668 2.0103993 2.4389560 0.7648271 1.5076055
#> Andrew Johnson         2.086668 2.0103993 2.4389560 0.7648271 1.5076055
#> Barack Obama           3.083916 0.1980421 0.2475526 3.1468759 3.2737465
#> Benjamin Harrison      2.086668 2.0103993 2.4389560 0.7648271 1.5076055
#> Calvin Coolidge        2.086668 2.0103993 2.4389560 0.7648271 1.5076055
#> Chester A. Arthur      2.086668 2.0103993 2.4389560 0.7648271 1.5076055
#> Donald J. Trump        3.083916 0.1980421 0.2475526 3.1468759 3.2737465
#> Dwight D. Eisenhower   3.083916 0.1980421 0.2475526 3.1468759 3.2737465
#> Franklin D. Roosevelt  3.083916 0.1980421 0.2475526 3.1468759 3.2737465
#> Franklin Pierce        2.086668 2.0103993 2.4389560 0.7648271 1.5076055
#> George Bush            3.083916 0.1980421 0.2475526 3.1468759 3.2737465
#> George W. Bush         3.083916 0.1980421 0.2475526 3.1468759 3.2737465
#> George Washington      2.086668 2.0103993 2.4389560 0.7648271 1.5076055
#> Gerald R. Ford         3.083916 0.1980421 0.2475526 3.1468759 3.2737465
#> Grover Cleveland       2.086668 2.0103993 2.4389560 0.7648271 1.5076055
#> Harry S. Truman        3.083916 0.1980421 0.2475526 3.1468759 3.2737465
#> Herbert Hoover         2.086668 2.0103993 2.4389560 0.7648271 1.5076055
#> James A. Garfield      2.086668 2.0103993 2.4389560 0.7648271 1.5076055
#> James Buchanan         2.086668 2.0103993 2.4389560 0.7648271 1.5076055
#> James K. Polk          2.086668 2.0103993 2.4389560 0.7648271 1.5076055
#> James Madison          2.086668 2.0103993 2.4389560 0.7648271 1.5076055
#> James Monroe           2.086668 2.0103993 2.4389560 0.7648271 1.5076055
#> Jimmy Carter           3.083916 0.1980421 0.2475526 3.1468759 3.2737465
#> John Adams             2.086668 2.0103993 2.4389560 0.7648271 1.5076055
#> John F. Kennedy        3.083916 0.1980421 0.2475526 3.1468759 3.2737465
#> John Quincy Adams      2.086668 2.0103993 2.4389560 0.7648271 1.5076055
#> John Tyler             2.086668 2.0103993 2.4389560 0.7648271 1.5076055
#> Lyndon B. Johnson      3.083916 0.1980421 0.2475526 3.1468759 3.2737465
#> Martin van Buren       2.086668 2.0103993 2.4389560 0.7648271 1.5076055
#> Millard Fillmore       2.086668 2.0103993 2.4389560 0.7648271 1.5076055
#> Richard Nixon          3.083916 0.1980421 0.2475526 3.1468759 3.2737465
#> Ronald Reagan          3.083916 0.1980421 0.2475526 3.1468759 3.2737465
#> Rutherford B. Hayes    2.086668 2.0103993 2.4389560 0.7648271 1.5076055
#> Theodore Roosevelt     2.086668 2.0103993 2.4389560 0.7648271 1.5076055
#> Thomas Jefferson       2.086668 2.0103993 2.4389560 0.7648271 1.5076055
#> Ulysses S. Grant       2.086668 2.0103993 2.4389560 0.7648271 1.5076055
#> Warren G. Harding      1.945910 0.0000000 0.6931472 1.0986123 0.6931472
#> William Henry Harrison 2.086668 2.0103993 2.4389560 0.7648271 1.5076055
#> William Howard Taft    2.086668 2.0103993 2.4389560 0.7648271 1.5076055
#> William J. Clinton     3.083916 0.1980421 0.2475526 3.1468759 3.2737465
#> William McKinley       2.086668 2.0103993 2.4389560 0.7648271 1.5076055
#> Woodrow Wilson         2.086668 2.0103993 2.4389560 0.7648271 1.5076055
#> Zachary Taylor         2.086668 2.0103993 2.4389560 0.7648271 1.5076055
#>                          econom     women     educ   school      bill    tariff
#> Abraham Lincoln        1.609623 0.5746917 1.892278 1.291019 1.9113003 1.9128817
#> Andrew Jackson         1.609623 0.5746917 1.892278 1.291019 1.9113003 1.9128817
#> Andrew Johnson         1.609623 0.5746917 1.892278 1.291019 1.9113003 1.9128817
#> Barack Obama           3.999372 2.9420389 3.463114 3.235803 2.8435069 0.8874667
#> Benjamin Harrison      1.609623 0.5746917 1.892278 1.291019 1.9113003 1.9128817
#> Calvin Coolidge        1.609623 0.5746917 1.892278 1.291019 1.9113003 1.9128817
#> Chester A. Arthur      1.609623 0.5746917 1.892278 1.291019 1.9113003 1.9128817
#> Donald J. Trump        3.999372 2.9420389 3.463114 3.235803 2.8435069 0.8874667
#> Dwight D. Eisenhower   3.999372 2.9420389 3.463114 3.235803 2.8435069 0.8874667
#> Franklin D. Roosevelt  3.999372 2.9420389 3.463114 3.235803 2.8435069 0.8874667
#> Franklin Pierce        1.609623 0.5746917 1.892278 1.291019 1.9113003 1.9128817
#> George Bush            3.999372 2.9420389 3.463114 3.235803 2.8435069 0.8874667
#> George W. Bush         3.999372 2.9420389 3.463114 3.235803 2.8435069 0.8874667
#> George Washington      1.609623 0.5746917 1.892278 1.291019 1.9113003 1.9128817
#> Gerald R. Ford         3.999372 2.9420389 3.463114 3.235803 2.8435069 0.8874667
#> Grover Cleveland       1.609623 0.5746917 1.892278 1.291019 1.9113003 1.9128817
#> Harry S. Truman        3.999372 2.9420389 3.463114 3.235803 2.8435069 0.8874667
#> Herbert Hoover         1.609623 0.5746917 1.892278 1.291019 1.9113003 1.9128817
#> James A. Garfield      1.609623 0.5746917 1.892278 1.291019 1.9113003 1.9128817
#> James Buchanan         1.609623 0.5746917 1.892278 1.291019 1.9113003 1.9128817
#> James K. Polk          1.609623 0.5746917 1.892278 1.291019 1.9113003 1.9128817
#> James Madison          1.609623 0.5746917 1.892278 1.291019 1.9113003 1.9128817
#> James Monroe           1.609623 0.5746917 1.892278 1.291019 1.9113003 1.9128817
#> Jimmy Carter           3.999372 2.9420389 3.463114 3.235803 2.8435069 0.8874667
#> John Adams             1.609623 0.5746917 1.892278 1.291019 1.9113003 1.9128817
#> John F. Kennedy        3.999372 2.9420389 3.463114 3.235803 2.8435069 0.8874667
#> John Quincy Adams      1.609623 0.5746917 1.892278 1.291019 1.9113003 1.9128817
#> John Tyler             1.609623 0.5746917 1.892278 1.291019 1.9113003 1.9128817
#> Lyndon B. Johnson      3.999372 2.9420389 3.463114 3.235803 2.8435069 0.8874667
#> Martin van Buren       1.609623 0.5746917 1.892278 1.291019 1.9113003 1.9128817
#> Millard Fillmore       1.609623 0.5746917 1.892278 1.291019 1.9113003 1.9128817
#> Richard Nixon          3.999372 2.9420389 3.463114 3.235803 2.8435069 0.8874667
#> Ronald Reagan          3.999372 2.9420389 3.463114 3.235803 2.8435069 0.8874667
#> Rutherford B. Hayes    1.609623 0.5746917 1.892278 1.291019 1.9113003 1.9128817
#> Theodore Roosevelt     1.609623 0.5746917 1.892278 1.291019 1.9113003 1.9128817
#> Thomas Jefferson       1.609623 0.5746917 1.892278 1.291019 1.9113003 1.9128817
#> Ulysses S. Grant       1.609623 0.5746917 1.892278 1.291019 1.9113003 1.9128817
#> Warren G. Harding      2.708050 1.0986123 2.197225 1.945910 0.6931472 2.8332133
#> William Henry Harrison 1.609623 0.5746917 1.892278 1.291019 1.9113003 1.9128817
#> William Howard Taft    1.609623 0.5746917 1.892278 1.291019 1.9113003 1.9128817
#> William J. Clinton     3.999372 2.9420389 3.463114 3.235803 2.8435069 0.8874667
#> William McKinley       1.609623 0.5746917 1.892278 1.291019 1.9113003 1.9128817
#> Woodrow Wilson         1.609623 0.5746917 1.892278 1.291019 1.9113003 1.9128817
#> Zachary Taylor         1.609623 0.5746917 1.892278 1.291019 1.9113003 1.9128817
#>                         unemploy    weapon       get   method   farmer
#> Abraham Lincoln        0.3871545 0.2768995 0.5113785 1.514215 1.093850
#> Andrew Jackson         0.3871545 0.2768995 0.5113785 1.514215 1.093850
#> Andrew Johnson         0.3871545 0.2768995 0.5113785 1.514215 1.093850
#> Barack Obama           2.7956399 2.7877462 3.0255971 1.461203 2.349569
#> Benjamin Harrison      0.3871545 0.2768995 0.5113785 1.514215 1.093850
#> Calvin Coolidge        0.3871545 0.2768995 0.5113785 1.514215 1.093850
#> Chester A. Arthur      0.3871545 0.2768995 0.5113785 1.514215 1.093850
#> Donald J. Trump        2.7956399 2.7877462 3.0255971 1.461203 2.349569
#> Dwight D. Eisenhower   2.7956399 2.7877462 3.0255971 1.461203 2.349569
#> Franklin D. Roosevelt  2.7956399 2.7877462 3.0255971 1.461203 2.349569
#> Franklin Pierce        0.3871545 0.2768995 0.5113785 1.514215 1.093850
#> George Bush            2.7956399 2.7877462 3.0255971 1.461203 2.349569
#> George W. Bush         2.7956399 2.7877462 3.0255971 1.461203 2.349569
#> George Washington      0.3871545 0.2768995 0.5113785 1.514215 1.093850
#> Gerald R. Ford         2.7956399 2.7877462 3.0255971 1.461203 2.349569
#> Grover Cleveland       0.3871545 0.2768995 0.5113785 1.514215 1.093850
#> Harry S. Truman        2.7956399 2.7877462 3.0255971 1.461203 2.349569
#> Herbert Hoover         0.3871545 0.2768995 0.5113785 1.514215 1.093850
#> James A. Garfield      0.3871545 0.2768995 0.5113785 1.514215 1.093850
#> James Buchanan         0.3871545 0.2768995 0.5113785 1.514215 1.093850
#> James K. Polk          0.3871545 0.2768995 0.5113785 1.514215 1.093850
#> James Madison          0.3871545 0.2768995 0.5113785 1.514215 1.093850
#> James Monroe           0.3871545 0.2768995 0.5113785 1.514215 1.093850
#> Jimmy Carter           2.7956399 2.7877462 3.0255971 1.461203 2.349569
#> John Adams             0.3871545 0.2768995 0.5113785 1.514215 1.093850
#> John F. Kennedy        2.7956399 2.7877462 3.0255971 1.461203 2.349569
#> John Quincy Adams      0.3871545 0.2768995 0.5113785 1.514215 1.093850
#> John Tyler             0.3871545 0.2768995 0.5113785 1.514215 1.093850
#> Lyndon B. Johnson      2.7956399 2.7877462 3.0255971 1.461203 2.349569
#> Martin van Buren       0.3871545 0.2768995 0.5113785 1.514215 1.093850
#> Millard Fillmore       0.3871545 0.2768995 0.5113785 1.514215 1.093850
#> Richard Nixon          2.7956399 2.7877462 3.0255971 1.461203 2.349569
#> Ronald Reagan          2.7956399 2.7877462 3.0255971 1.461203 2.349569
#> Rutherford B. Hayes    0.3871545 0.2768995 0.5113785 1.514215 1.093850
#> Theodore Roosevelt     0.3871545 0.2768995 0.5113785 1.514215 1.093850
#> Thomas Jefferson       0.3871545 0.2768995 0.5113785 1.514215 1.093850
#> Ulysses S. Grant       0.3871545 0.2768995 0.5113785 1.514215 1.093850
#> Warren G. Harding      1.6094379 0.0000000 0.6931472 1.386294 2.302585
#> William Henry Harrison 0.3871545 0.2768995 0.5113785 1.514215 1.093850
#> William Howard Taft    0.3871545 0.2768995 0.5113785 1.514215 1.093850
#> William J. Clinton     2.7956399 2.7877462 3.0255971 1.461203 2.349569
#> William McKinley       0.3871545 0.2768995 0.5113785 1.514215 1.093850
#> Woodrow Wilson         0.3871545 0.2768995 0.5113785 1.514215 1.093850
#> Zachary Taylor         0.3871545 0.2768995 0.5113785 1.514215 1.093850
#>                         challeng   achiev democraci      area   billion
#> Abraham Lincoln        0.3427457 1.049185 0.2999833 0.9618479 0.2441388
#> Andrew Jackson         0.3427457 1.049185 0.2999833 0.9618479 0.2441388
#> Andrew Johnson         0.3427457 1.049185 0.2999833 0.9618479 0.2441388
#> Barack Obama           3.1210443 3.403615 2.5199520 3.0214103 3.7429589
#> Benjamin Harrison      0.3427457 1.049185 0.2999833 0.9618479 0.2441388
#> Calvin Coolidge        0.3427457 1.049185 0.2999833 0.9618479 0.2441388
#> Chester A. Arthur      0.3427457 1.049185 0.2999833 0.9618479 0.2441388
#> Donald J. Trump        3.1210443 3.403615 2.5199520 3.0214103 3.7429589
#> Dwight D. Eisenhower   3.1210443 3.403615 2.5199520 3.0214103 3.7429589
#> Franklin D. Roosevelt  3.1210443 3.403615 2.5199520 3.0214103 3.7429589
#> Franklin Pierce        0.3427457 1.049185 0.2999833 0.9618479 0.2441388
#> George Bush            3.1210443 3.403615 2.5199520 3.0214103 3.7429589
#> George W. Bush         3.1210443 3.403615 2.5199520 3.0214103 3.7429589
#> George Washington      0.3427457 1.049185 0.2999833 0.9618479 0.2441388
#> Gerald R. Ford         3.1210443 3.403615 2.5199520 3.0214103 3.7429589
#> Grover Cleveland       0.3427457 1.049185 0.2999833 0.9618479 0.2441388
#> Harry S. Truman        3.1210443 3.403615 2.5199520 3.0214103 3.7429589
#> Herbert Hoover         0.3427457 1.049185 0.2999833 0.9618479 0.2441388
#> James A. Garfield      0.3427457 1.049185 0.2999833 0.9618479 0.2441388
#> James Buchanan         0.3427457 1.049185 0.2999833 0.9618479 0.2441388
#> James K. Polk          0.3427457 1.049185 0.2999833 0.9618479 0.2441388
#> James Madison          0.3427457 1.049185 0.2999833 0.9618479 0.2441388
#> James Monroe           0.3427457 1.049185 0.2999833 0.9618479 0.2441388
#> Jimmy Carter           3.1210443 3.403615 2.5199520 3.0214103 3.7429589
#> John Adams             0.3427457 1.049185 0.2999833 0.9618479 0.2441388
#> John F. Kennedy        3.1210443 3.403615 2.5199520 3.0214103 3.7429589
#> John Quincy Adams      0.3427457 1.049185 0.2999833 0.9618479 0.2441388
#> John Tyler             0.3427457 1.049185 0.2999833 0.9618479 0.2441388
#> Lyndon B. Johnson      3.1210443 3.403615 2.5199520 3.0214103 3.7429589
#> Martin van Buren       0.3427457 1.049185 0.2999833 0.9618479 0.2441388
#> Millard Fillmore       0.3427457 1.049185 0.2999833 0.9618479 0.2441388
#> Richard Nixon          3.1210443 3.403615 2.5199520 3.0214103 3.7429589
#> Ronald Reagan          3.1210443 3.403615 2.5199520 3.0214103 3.7429589
#> Rutherford B. Hayes    0.3427457 1.049185 0.2999833 0.9618479 0.2441388
#> Theodore Roosevelt     0.3427457 1.049185 0.2999833 0.9618479 0.2441388
#> Thomas Jefferson       0.3427457 1.049185 0.2999833 0.9618479 0.2441388
#> Ulysses S. Grant       0.3427457 1.049185 0.2999833 0.9618479 0.2441388
#> Warren G. Harding      1.6094379 1.386294 1.6094379 1.0986123 1.3862944
#> William Henry Harrison 0.3427457 1.049185 0.2999833 0.9618479 0.2441388
#> William Howard Taft    0.3427457 1.049185 0.2999833 0.9618479 0.2441388
#> William J. Clinton     3.1210443 3.403615 2.5199520 3.0214103 3.7429589
#> William McKinley       0.3427457 1.049185 0.2999833 0.9618479 0.2441388
#> Woodrow Wilson         0.3427457 1.049185 0.2999833 0.9618479 0.2441388
#> Zachary Taylor         0.3427457 1.049185 0.2999833 0.9618479 0.2441388
#>                         problem     basic   budget      goal       job
#> Abraham Lincoln        1.162271 0.1109957 0.430126 0.1033011 0.1474712
#> Andrew Jackson         1.162271 0.1109957 0.430126 0.1033011 0.1474712
#> Andrew Johnson         1.162271 0.1109957 0.430126 0.1033011 0.1474712
#> Barack Obama           3.659555 2.8387802 3.795996 3.2118895 3.9587172
#> Benjamin Harrison      1.162271 0.1109957 0.430126 0.1033011 0.1474712
#> Calvin Coolidge        1.162271 0.1109957 0.430126 0.1033011 0.1474712
#> Chester A. Arthur      1.162271 0.1109957 0.430126 0.1033011 0.1474712
#> Donald J. Trump        3.659555 2.8387802 3.795996 3.2118895 3.9587172
#> Dwight D. Eisenhower   3.659555 2.8387802 3.795996 3.2118895 3.9587172
#> Franklin D. Roosevelt  3.659555 2.8387802 3.795996 3.2118895 3.9587172
#> Franklin Pierce        1.162271 0.1109957 0.430126 0.1033011 0.1474712
#> George Bush            3.659555 2.8387802 3.795996 3.2118895 3.9587172
#> George W. Bush         3.659555 2.8387802 3.795996 3.2118895 3.9587172
#> George Washington      1.162271 0.1109957 0.430126 0.1033011 0.1474712
#> Gerald R. Ford         3.659555 2.8387802 3.795996 3.2118895 3.9587172
#> Grover Cleveland       1.162271 0.1109957 0.430126 0.1033011 0.1474712
#> Harry S. Truman        3.659555 2.8387802 3.795996 3.2118895 3.9587172
#> Herbert Hoover         1.162271 0.1109957 0.430126 0.1033011 0.1474712
#> James A. Garfield      1.162271 0.1109957 0.430126 0.1033011 0.1474712
#> James Buchanan         1.162271 0.1109957 0.430126 0.1033011 0.1474712
#> James K. Polk          1.162271 0.1109957 0.430126 0.1033011 0.1474712
#> James Madison          1.162271 0.1109957 0.430126 0.1033011 0.1474712
#> James Monroe           1.162271 0.1109957 0.430126 0.1033011 0.1474712
#> Jimmy Carter           3.659555 2.8387802 3.795996 3.2118895 3.9587172
#> John Adams             1.162271 0.1109957 0.430126 0.1033011 0.1474712
#> John F. Kennedy        3.659555 2.8387802 3.795996 3.2118895 3.9587172
#> John Quincy Adams      1.162271 0.1109957 0.430126 0.1033011 0.1474712
#> John Tyler             1.162271 0.1109957 0.430126 0.1033011 0.1474712
#> Lyndon B. Johnson      3.659555 2.8387802 3.795996 3.2118895 3.9587172
#> Martin van Buren       1.162271 0.1109957 0.430126 0.1033011 0.1474712
#> Millard Fillmore       1.162271 0.1109957 0.430126 0.1033011 0.1474712
#> Richard Nixon          3.659555 2.8387802 3.795996 3.2118895 3.9587172
#> Ronald Reagan          3.659555 2.8387802 3.795996 3.2118895 3.9587172
#> Rutherford B. Hayes    1.162271 0.1109957 0.430126 0.1033011 0.1474712
#> Theodore Roosevelt     1.162271 0.1109957 0.430126 0.1033011 0.1474712
#> Thomas Jefferson       1.162271 0.1109957 0.430126 0.1033011 0.1474712
#> Ulysses S. Grant       1.162271 0.1109957 0.430126 0.1033011 0.1474712
#> Warren G. Harding      3.526361 1.0986123 1.386294 0.6931472 0.0000000
#> William Henry Harrison 1.162271 0.1109957 0.430126 0.1033011 0.1474712
#> William Howard Taft    1.162271 0.1109957 0.430126 0.1033011 0.1474712
#> William J. Clinton     3.659555 2.8387802 3.795996 3.2118895 3.9587172
#> William McKinley       1.162271 0.1109957 0.430126 0.1033011 0.1474712
#> Woodrow Wilson         1.162271 0.1109957 0.430126 0.1033011 0.1474712
#> Zachary Taylor         1.162271 0.1109957 0.430126 0.1033011 0.1474712
#>                            level  nuclear   percent   program     spend
#> Abraham Lincoln        0.5037702 0.000000 0.1527868 0.4188332 0.2746740
#> Andrew Jackson         0.5037702 0.000000 0.1527868 0.4188332 0.2746740
#> Andrew Johnson         0.5037702 0.000000 0.1527868 0.4188332 0.2746740
#> Barack Obama           3.0683132 2.559875 3.4095268 4.4328005 3.1180550
#> Benjamin Harrison      0.5037702 0.000000 0.1527868 0.4188332 0.2746740
#> Calvin Coolidge        0.5037702 0.000000 0.1527868 0.4188332 0.2746740
#> Chester A. Arthur      0.5037702 0.000000 0.1527868 0.4188332 0.2746740
#> Donald J. Trump        3.0683132 2.559875 3.4095268 4.4328005 3.1180550
#> Dwight D. Eisenhower   3.0683132 2.559875 3.4095268 4.4328005 3.1180550
#> Franklin D. Roosevelt  3.0683132 2.559875 3.4095268 4.4328005 3.1180550
#> Franklin Pierce        0.5037702 0.000000 0.1527868 0.4188332 0.2746740
#> George Bush            3.0683132 2.559875 3.4095268 4.4328005 3.1180550
#> George W. Bush         3.0683132 2.559875 3.4095268 4.4328005 3.1180550
#> George Washington      0.5037702 0.000000 0.1527868 0.4188332 0.2746740
#> Gerald R. Ford         3.0683132 2.559875 3.4095268 4.4328005 3.1180550
#> Grover Cleveland       0.5037702 0.000000 0.1527868 0.4188332 0.2746740
#> Harry S. Truman        3.0683132 2.559875 3.4095268 4.4328005 3.1180550
#> Herbert Hoover         0.5037702 0.000000 0.1527868 0.4188332 0.2746740
#> James A. Garfield      0.5037702 0.000000 0.1527868 0.4188332 0.2746740
#> James Buchanan         0.5037702 0.000000 0.1527868 0.4188332 0.2746740
#> James K. Polk          0.5037702 0.000000 0.1527868 0.4188332 0.2746740
#> James Madison          0.5037702 0.000000 0.1527868 0.4188332 0.2746740
#> James Monroe           0.5037702 0.000000 0.1527868 0.4188332 0.2746740
#> Jimmy Carter           3.0683132 2.559875 3.4095268 4.4328005 3.1180550
#> John Adams             0.5037702 0.000000 0.1527868 0.4188332 0.2746740
#> John F. Kennedy        3.0683132 2.559875 3.4095268 4.4328005 3.1180550
#> John Quincy Adams      0.5037702 0.000000 0.1527868 0.4188332 0.2746740
#> John Tyler             0.5037702 0.000000 0.1527868 0.4188332 0.2746740
#> Lyndon B. Johnson      3.0683132 2.559875 3.4095268 4.4328005 3.1180550
#> Martin van Buren       0.5037702 0.000000 0.1527868 0.4188332 0.2746740
#> Millard Fillmore       0.5037702 0.000000 0.1527868 0.4188332 0.2746740
#> Richard Nixon          3.0683132 2.559875 3.4095268 4.4328005 3.1180550
#> Ronald Reagan          3.0683132 2.559875 3.4095268 4.4328005 3.1180550
#> Rutherford B. Hayes    0.5037702 0.000000 0.1527868 0.4188332 0.2746740
#> Theodore Roosevelt     0.5037702 0.000000 0.1527868 0.4188332 0.2746740
#> Thomas Jefferson       0.5037702 0.000000 0.1527868 0.4188332 0.2746740
#> Ulysses S. Grant       0.5037702 0.000000 0.1527868 0.4188332 0.2746740
#> Warren G. Harding      1.7917595 0.000000 0.0000000 2.1972246 0.6931472
#> William Henry Harrison 0.5037702 0.000000 0.1527868 0.4188332 0.2746740
#> William Howard Taft    0.5037702 0.000000 0.1527868 0.4188332 0.2746740
#> William J. Clinton     3.0683132 2.559875 3.4095268 4.4328005 3.1180550
#> William McKinley       0.5037702 0.000000 0.1527868 0.4188332 0.2746740
#> Woodrow Wilson         0.5037702 0.000000 0.1527868 0.4188332 0.2746740
#> Zachary Taylor         0.5037702 0.000000 0.1527868 0.4188332 0.2746740
#>                         technolog     today  tonight    worker    inflat
#> Abraham Lincoln        0.02390163 0.2544055 0.000000 0.2961965 0.3494689
#> Andrew Jackson         0.02390163 0.2544055 0.000000 0.2961965 0.3494689
#> Andrew Johnson         0.02390163 0.2544055 0.000000 0.2961965 0.3494689
#> Barack Obama           2.68228351 3.7233467 2.596381 3.1347856 2.5319042
#> Benjamin Harrison      0.02390163 0.2544055 0.000000 0.2961965 0.3494689
#> Calvin Coolidge        0.02390163 0.2544055 0.000000 0.2961965 0.3494689
#> Chester A. Arthur      0.02390163 0.2544055 0.000000 0.2961965 0.3494689
#> Donald J. Trump        2.68228351 3.7233467 2.596381 3.1347856 2.5319042
#> Dwight D. Eisenhower   2.68228351 3.7233467 2.596381 3.1347856 2.5319042
#> Franklin D. Roosevelt  2.68228351 3.7233467 2.596381 3.1347856 2.5319042
#> Franklin Pierce        0.02390163 0.2544055 0.000000 0.2961965 0.3494689
#> George Bush            2.68228351 3.7233467 2.596381 3.1347856 2.5319042
#> George W. Bush         2.68228351 3.7233467 2.596381 3.1347856 2.5319042
#> George Washington      0.02390163 0.2544055 0.000000 0.2961965 0.3494689
#> Gerald R. Ford         2.68228351 3.7233467 2.596381 3.1347856 2.5319042
#> Grover Cleveland       0.02390163 0.2544055 0.000000 0.2961965 0.3494689
#> Harry S. Truman        2.68228351 3.7233467 2.596381 3.1347856 2.5319042
#> Herbert Hoover         0.02390163 0.2544055 0.000000 0.2961965 0.3494689
#> James A. Garfield      0.02390163 0.2544055 0.000000 0.2961965 0.3494689
#> James Buchanan         0.02390163 0.2544055 0.000000 0.2961965 0.3494689
#> James K. Polk          0.02390163 0.2544055 0.000000 0.2961965 0.3494689
#> James Madison          0.02390163 0.2544055 0.000000 0.2961965 0.3494689
#> James Monroe           0.02390163 0.2544055 0.000000 0.2961965 0.3494689
#> Jimmy Carter           2.68228351 3.7233467 2.596381 3.1347856 2.5319042
#> John Adams             0.02390163 0.2544055 0.000000 0.2961965 0.3494689
#> John F. Kennedy        2.68228351 3.7233467 2.596381 3.1347856 2.5319042
#> John Quincy Adams      0.02390163 0.2544055 0.000000 0.2961965 0.3494689
#> John Tyler             0.02390163 0.2544055 0.000000 0.2961965 0.3494689
#> Lyndon B. Johnson      2.68228351 3.7233467 2.596381 3.1347856 2.5319042
#> Martin van Buren       0.02390163 0.2544055 0.000000 0.2961965 0.3494689
#> Millard Fillmore       0.02390163 0.2544055 0.000000 0.2961965 0.3494689
#> Richard Nixon          2.68228351 3.7233467 2.596381 3.1347856 2.5319042
#> Ronald Reagan          2.68228351 3.7233467 2.596381 3.1347856 2.5319042
#> Rutherford B. Hayes    0.02390163 0.2544055 0.000000 0.2961965 0.3494689
#> Theodore Roosevelt     0.02390163 0.2544055 0.000000 0.2961965 0.3494689
#> Thomas Jefferson       0.02390163 0.2544055 0.000000 0.2961965 0.3494689
#> Ulysses S. Grant       0.02390163 0.2544055 0.000000 0.2961965 0.3494689
#> Warren G. Harding      0.00000000 2.3978953 0.000000 0.6931472 1.6094379
#> William Henry Harrison 0.02390163 0.2544055 0.000000 0.2961965 0.3494689
#> William Howard Taft    0.02390163 0.2544055 0.000000 0.2961965 0.3494689
#> William J. Clinton     2.68228351 3.7233467 2.596381 3.1347856 2.5319042
#> William McKinley       0.02390163 0.2544055 0.000000 0.2961965 0.3494689
#> Woodrow Wilson         0.02390163 0.2544055 0.000000 0.2961965 0.3494689
#> Zachary Taylor         0.02390163 0.2544055 0.000000 0.2961965 0.3494689
#>                          soviet
#> Abraham Lincoln        0.000000
#> Andrew Jackson         0.000000
#> Andrew Johnson         0.000000
#> Barack Obama           2.493221
#> Benjamin Harrison      0.000000
#> Calvin Coolidge        0.000000
#> Chester A. Arthur      0.000000
#> Donald J. Trump        2.493221
#> Dwight D. Eisenhower   2.493221
#> Franklin D. Roosevelt  2.493221
#> Franklin Pierce        0.000000
#> George Bush            2.493221
#> George W. Bush         2.493221
#> George Washington      0.000000
#> Gerald R. Ford         2.493221
#> Grover Cleveland       0.000000
#> Harry S. Truman        2.493221
#> Herbert Hoover         0.000000
#> James A. Garfield      0.000000
#> James Buchanan         0.000000
#> James K. Polk          0.000000
#> James Madison          0.000000
#> James Monroe           0.000000
#> Jimmy Carter           2.493221
#> John Adams             0.000000
#> John F. Kennedy        2.493221
#> John Quincy Adams      0.000000
#> John Tyler             0.000000
#> Lyndon B. Johnson      2.493221
#> Martin van Buren       0.000000
#> Millard Fillmore       0.000000
#> Richard Nixon          2.493221
#> Ronald Reagan          2.493221
#> Rutherford B. Hayes    0.000000
#> Theodore Roosevelt     0.000000
#> Thomas Jefferson       0.000000
#> Ulysses S. Grant       0.000000
#> Warren G. Harding      0.000000
#> William Henry Harrison 0.000000
#> William Howard Taft    0.000000
#> William J. Clinton     2.493221
#> William McKinley       0.000000
#> Woodrow Wilson         0.000000
#> Zachary Taylor         0.000000