Control for CARP fits

Set CARP algorithm parameters

carp.control(obs.labels = NULL, var.labels = NULL, X.center = TRUE,
  X.scale = FALSE, phi = NULL, rho = 1, weights = NULL, k = NULL,
  weight.dist = "euclidean", weight.dist.p = 2, ncores = 1L,
  max.iter = 1000000L, burn.in = 50L, alg.type = "carpviz",
  t = 1.05, npcs = NULL, dendrogram.scale = NULL, ...)

Arguments

obs.labels
A character vector of length \(n\): observations (row) labels

var.labels
A character vector of length \(p\): variable (column) labels

X.center
A logical: Should X be centered columnwise?

X.scale
A logical: Should X be scaled columnwise?

phi
A positive real number: the scale factor used in the RBF kernel

rho
For advanced users only (not advisable to change): the penalty parameter used for the augmented Lagrangian.

weights
A vector of positive number of length choose(n,2).

k
An positive integer: the number of neighbors used to create sparse weights

weight.dist
A string indicating the distance metric used to calculate weights. See distance for details.

weight.dist.p
The exponent used to calculate the Minkowski distance if weight.dist = "minkowski". See distance for details.

ncores
An positive integer: the number of cores to use.

max.iter
An integer: the maximum number of CARP iterations.

burn.in
An integer: the number of initial iterations at a fixed (small) value of \(\lambda\)

alg.type
Which CARP variant to use. Allowed values are

"carp" - The standard CARP algorithm with \(L2\) penalty;

"carpviz" - The back-tracking CARP algorithm with \(L2\) penalty;

"carpl1" - The standard CARP algorithm with \(L1\) penalty; and

"carpvizl1" - The back-tracking CARP algorithm with \(L1\) penalty.

t
A number greater than 1: the size of the multiplicative update to the cluster fusion regularization parameter (not used by back-tracking variants). Typically on the scale of 1.005 to 1.1.

npcs
An integer >= 2. The number of principal components to compute for path visualization.

dendrogram.scale
A character string denoting how the scale of dendrogram regularization proportions should be visualized. Choices are 'original' or 'log'; if not provided, a data-driven heuristic choice is used.

...
Unused arguements. An error will be thrown if any unrecognized arguments as given.

Value

A list containing the CARP algorithm parameters.

Details

This function constructs a list containing additional arguments to control the behavior of the CARP algorithm. It is typically only used internally by CARP, but may be useful to advanced users who wish to construct the control argument directly.