ClustRViz Package Logging Functionality

Control the global logging level for the clustRviz package.

clustRviz_logger_level(
  level = c("ERROR", "WARNING", "MESSAGE", "INFO", "DEBUG")
)

Arguments

level
The desired new log level. Available levels are

ERROR - corresponding to base::stop;

WARNING - corresponding to base::warning;

MESSAGE - corresponding to base::message;

INFO; and

DEBUG.

If omitted, the log level is not changed (and the current level is still returned invisibly.) See below for details about the different levels.

Value

The previous log level (invisibly).

Details

The clustRviz package has a multi-level logging system, with a single global log level; (which applies to both R and C++ level functionality.) the levels are, in decreasing order, ERROR, WARNING, MESSAGE (default), INFO, DEBUG.

To change the amount of output from the clustRviz package, the clustRviz_logger_level function can be used to adjust the global log level. The INFO and DEBUG levels can be quite verbose and may significantly slow down the package.

Examples

# Switch to INFO level and fit somewhat loudly
clustRviz_logger_level("INFO")
CARP(presidential_speech)

#> Pre-computing weights and edge sets

#> Computing Convex Clustering [CARP] Path

#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 1
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 2
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 3
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 4
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 5
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 6
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 7
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 8
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 9
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 10
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 11
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 12
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 13
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 14
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 15
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 16
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 17
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 18
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 19
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 20
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 21
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 22
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 23
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 24
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 25
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 26
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 27
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 28
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 29
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 30
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 31
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 32
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 33
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 34
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 35
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 36
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 37
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 38
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 39
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 40
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 41
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 42
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 43
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 44
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 45
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 46
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 47
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 48
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 49
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 50
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 51
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 52
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 53
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 54
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 55
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 56
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 57
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 58
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 59
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 60
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 61
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 62
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 63
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 64
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 65
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 66
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 67
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 68
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 69
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 70
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 71
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 72
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 73
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 74
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 75
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 76
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 77
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 78
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 79
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 80
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 81
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 82
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 83
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 84
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 85
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 86
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 87
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 88
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 89
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 90
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 91
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 92
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 93
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 94
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 95
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 96
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 97
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 98
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 99
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 100
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 101
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 102
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 103
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 104
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 105
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 106
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 107
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 108
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 109
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 110
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 111
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 112
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 113
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 114
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 115
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 116
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 117
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 118
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 119
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 120
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 121
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 122
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 123
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 124
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 125
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 126
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 127
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 128
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 129
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 130
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 131
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 132
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 133
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 134
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 135
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 136
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 137
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 138
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 139
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 140
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 141
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 142
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 143
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 144
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 145
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 146
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 147
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 148
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 149
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 150
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 151
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 152
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 153
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 154
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 155
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 156
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 157
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 158
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 159
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 160
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 161
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 162
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 163
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 164
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 165
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 166
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 167
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 168
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 169
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 170
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 171
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 172
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 173
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 174
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 175
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 176
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 177
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 178
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 179
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 180
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 181
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 182
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 183
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 184
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 185
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 186
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 187
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 188
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 189
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 190
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 191
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 192
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 193
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 194
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 195
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 196
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 197
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 198
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 199
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 200
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 201
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 202
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 203
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 204
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 205
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 206
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 207
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 208
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 209
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 210
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 211
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 212
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 213
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 214
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 215
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 216
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 217
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 218
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 219
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 220
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 221
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 222
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 223
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 224
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 225
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 226
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 227
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 228
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 229
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 230
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 231
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 232
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 233
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 234
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 235
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 236
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 237
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 238
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 239
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 240
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 241
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 242
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 243
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 244
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 245
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 246
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 247
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 248
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 249
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 250
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 251
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 252
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 253
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 254
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 255
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 256
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 257
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 258
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 259
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 260
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 261
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 262
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 263
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 264
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 265
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 266
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 267
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 268
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 269
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 270
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 271
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 272
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 273
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 274
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 275
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 276
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 277
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 278
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 279
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 280
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 281
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 282
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 283
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 284
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 285
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 286
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 287
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 288
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 289
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 290
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 291
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 292
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 293
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 294
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 295
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 296
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 297
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 298
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 299
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 300
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 301
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 302
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 303
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 304
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 305
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 306
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 307
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 308
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 309
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 310
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 311
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 312
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 313
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 314
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 315
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 316
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 317
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 318
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 319
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 320
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 321
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 322
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 323
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 324
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 325
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 326
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 327
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 328
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 329
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 330
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 331
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 332
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 333
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 334
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 335
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 336
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 337
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 338
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 339
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 340
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 341
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 342
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 343
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 344
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 345
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 346
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 347
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 348
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 349
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 350
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 351
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 352
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 353
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 354
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 355
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 356
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 357
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 358
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 359
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 360
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 361
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 362
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 363
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 364
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 365
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 366
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 367
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 368
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 369
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 370
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 371
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 372
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 373
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 374
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 375
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 376
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 377
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 378
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 379
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 380
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 381
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 382
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 383
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 384
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 385
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 386
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 387
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 388
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 389
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 390
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 391
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 392
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 393
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 394
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 395
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 396
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 397
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 398
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 399
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 400
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 401
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 402
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 403
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 404
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 405
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 406
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 407
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 408
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 409
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 410
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 411
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 412
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 413
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 414
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 415
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 416
#> [INFO] -- 2020-08-30 20:42:58 -- Resizing storage from 66 to 132 iterations.
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 417
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 418
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 419
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 420
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 421
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 422
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 423
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 424
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 425
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 426
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 427
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 428
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 429
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 430
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 431
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 432
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 433
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 434
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 435
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 436
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 437
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 438
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 439
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 440
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 441
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 442
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 443
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 444
#> [INFO] -- 2020-08-30 20:42:58 -- Beginning iteration k = 445

#> Post-processing

#> CARP Fit Summary
#> ====================
#> 
#> Algorithm: CARP (t = 1.05) 
#> Fit Time: 0.446 secs 
#> Total Time: 0.839 secs 
#> 
#> Number of Observations: 44 
#> Number of Variables:    75 
#> 
#> Pre-processing options:
#>  - Columnwise centering: TRUE 
#>  - Columnwise scaling:   FALSE 
#> 
#> Weights:
#>  - Source: Radial Basis Function Kernel Weights
#>  - Distance Metric: Euclidean
#>  - Scale parameter (phi): 0.01 [Data-Driven]
#>  - Sparsified: 4 Nearest Neighbors [Data-Driven]
#> 

# Return to default behavior
clustRviz_logger_level("MESSAGE")
# Will give a few messages about phases of fitting
CARP(presidential_speech)

#> Pre-computing weights and edge sets

#> Computing Convex Clustering [CARP] Path

#> Post-processing

#> CARP Fit Summary
#> ====================
#> 
#> Algorithm: CARP (t = 1.05) 
#> Fit Time: 0.150 secs 
#> Total Time: 0.953 secs 
#> 
#> Number of Observations: 44 
#> Number of Variables:    75 
#> 
#> Pre-processing options:
#>  - Columnwise centering: TRUE 
#>  - Columnwise scaling:   FALSE 
#> 
#> Weights:
#>  - Source: Radial Basis Function Kernel Weights
#>  - Distance Metric: Euclidean
#>  - Scale parameter (phi): 0.01 [Data-Driven]
#>  - Sparsified: 4 Nearest Neighbors [Data-Driven]
#> 

# If we switch to ERRORs only, no progress message
clustRviz_logger_level("ERROR")
CARP(presidential_speech)

#> CARP Fit Summary
#> ====================
#> 
#> Algorithm: CARP (t = 1.05) 
#> Fit Time: 0.432 secs 
#> Total Time: 1.441 secs 
#> 
#> Number of Observations: 44 
#> Number of Variables:    75 
#> 
#> Pre-processing options:
#>  - Columnwise centering: TRUE 
#>  - Columnwise scaling:   FALSE 
#> 
#> Weights:
#>  - Source: Radial Basis Function Kernel Weights
#>  - Distance Metric: Euclidean
#>  - Scale parameter (phi): 0.01 [Data-Driven]
#>  - Sparsified: 4 Nearest Neighbors [Data-Driven]
#> 

# Return to default
clustRviz_logger_level("MESSAGE")

# Normally this prints messages as the paths are calculated
CARP(presidential_speech)

#> Pre-computing weights and edge sets

#> Computing Convex Clustering [CARP] Path

#> Post-processing

#> CARP Fit Summary
#> ====================
#> 
#> Algorithm: CARP (t = 1.05) 
#> Fit Time: 0.461 secs 
#> Total Time: 1.549 secs 
#> 
#> Number of Observations: 44 
#> Number of Variables:    75 
#> 
#> Pre-processing options:
#>  - Columnwise centering: TRUE 
#>  - Columnwise scaling:   FALSE 
#> 
#> Weights:
#>  - Source: Radial Basis Function Kernel Weights
#>  - Distance Metric: Euclidean
#>  - Scale parameter (phi): 0.01 [Data-Driven]
#>  - Sparsified: 4 Nearest Neighbors [Data-Driven]
#> 

# But it can be suppressed using standard R functions
suppressMessages(CARP(presidential_speech))

#> CARP Fit Summary
#> ====================
#> 
#> Algorithm: CARP (t = 1.05) 
#> Fit Time: 0.417 secs 
#> Total Time: 1.216 secs 
#> 
#> Number of Observations: 44 
#> Number of Variables:    75 
#> 
#> Pre-processing options:
#>  - Columnwise centering: TRUE 
#>  - Columnwise scaling:   FALSE 
#> 
#> Weights:
#>  - Source: Radial Basis Function Kernel Weights
#>  - Distance Metric: Euclidean
#>  - Scale parameter (phi): 0.01 [Data-Driven]
#>  - Sparsified: 4 Nearest Neighbors [Data-Driven]
#>