ClustRViz Options

Advanced control of algorithmic options for CARP and CBASS. The clustRviz_reset_options function returns options to "factory-fresh" settings.

clustRviz_options(...)

clustRviz_reset_options()

Arguments

...
Options (to be passed by name). See below for available options.

Details

The following options can be set by name:

epsilon The initial step size (fixed during the "burn-in" period)

stopping_threshold: Stopping threshold to use for convex_clustering, convex_biclustering, CARP(exact = TRUE), and CBASS(exact = TRUE). By default, this is set to 1e-10 - a very conservative threshold: making it larger can significantly improve performance

max_iter An integer: the maximum number of iterations to perform

max_inner_iter An integer: the maximum number of iterations for the iterative solvers to perform at a single value of \(\lambda\).

burn_in An integer: the number of initial iterations at a fixed (small) value of \(\gamma\)

viz_initial_step The initial (large) step size used in back-tracking (CARP-VIZ and CBASS-VIZ) algorithms.

viz_small_step The secondary (small) step size used in back-tracking (CARP-VIZ and CBASS-VIZ) algorithms.

viz_max_inner_iter The maximum number of iterations to perform in the inner loop of back-tracking (CARP-VIZ and CBASS-VIZ) algorithms.

keep CARP and CBASS keep every keep-th iteration even if no fusions are detected. Increasing this parameter may improve performance, at the expense of returning a finer grid.

rho For advanced users only (not advisable to change): the penalty parameter used for the augmented Lagrangian.

keep_debug_info: Should additional debug info (currently only the V-path) be kept?