Visualize the results of Convex Clustering (CARP)

plot.CARP provides a range of ways to visualize the results of convex clustering, including:

A dendrogram, illustrating the nested cluster hierarchy inferred from the convex clustering solution path (type = "dendrogram");

A static path plot, showing the coalescence of the estimated cluster centroids at a fixed value of the regularization parameter is increased (type = "path");

A gganimate plot, showing the coalescence of the estimated cluster centroids as the regularization parameter is increased (type = "dynamic_path"); and

A shiny app, which can display the clustering solutions as a "movie" or allow for interactive exploration (type = "interactive").

# S3 method for CARP
plot(
  x,
  ...,
  type = c("dendrogram", "path", "dynamic_path", "js", "interactive"),
  axis = c("PC1", "PC2"),
  dend.branch.width = 2,
  dend.labels.cex = 0.6,
  dend.ylab.cex = 1.2,
  percent,
  k,
  percent.seq = seq(0, 1, length.out = 100),
  max.nclust = 9,
  min.nclust = 1
)

saveviz(x, ...)

# S3 method for CARP
saveviz(
  x,
  file.name,
  type = c("dendrogram", "path", "js"),
  dynamic = TRUE,
  axis = c("PC1", "PC2"),
  dend.branch.width = 2,
  dend.labels.cex = 0.6,
  percent,
  k,
  percent.seq = seq(from = 0, to = 1, by = 0.01),
  width = 8,
  height = 5,
  units = c("in", "cm", "mm", "px"),
  ...
)

Arguments

x
An object of class CARP as returned by CARP

...
Additional arguments, which are handled differently for different values of type.

When type is "path", "dynamic_path", or "interactive", the presence of unknown arguments triggers an error;

when type == "dendrogram" ... is forwarded to plot.dendrogram; and

when type == "js", ... is forwarded to heatmaply.

See the documentation of the linked functions for details about additional supported arguments. saveviz passes arguments to the corresponding plot type.

type
A string indicating the type of visualization to show (see details above).

axis
A character vector of length two indicating which features or principal components to use as the axes in the type = "path" visualization. Currently only features like "PC1" or "PC2" (indicating the first principal component projections) are supported.

dend.branch.width
Branch width for dendrogram plots (ignored for other plot types) - must be positive.

dend.labels.cex
Label size for dendrogram plots (ignored for other plot types) - must be positive.

dend.ylab.cex
Y-axis size for dendrogram plots (ignored for other plot types) - must be positive.

percent
A number between 0 and 1, giving the regularization level (as a fraction of the final regularization level used) at which to assign clusters in the static (type = "dendrogram" or type = "path") plots.

k
An integer indicating the desired number of clusters to be displayed in the static plots. If no CARP iteration with exactly this many clusters is found, the first iterate with fewer than k clusters is used.

percent.seq
A grid of values of percent along which to generate dynamic visualizations (if dynamic == TRUE)

max.nclust
a positive integer. The maximum number of clusters to display in the interactive plot.

min.nclust
a positive value. The minimum number of clusters to display in the interactive plot.

file.name
The name of the output file. The type of resulting image is determined from the extension. If dynamic = TRUE, the extension is changed to .gif internally. If type = "js", the extension is changed to .html internally.

dynamic
Should the resulting animation be dynamic (animated) or not? If TRUE, a dynamic visualization which varies along the CARP solution path at a grid given by percent.seq is produced (as a GIF). If FALSE, a fixed visualization at a single solution (determined by either percent or k if supplied) is produced. Currently ignored when type = "js".

width
The width of the output, given in units

height
The height of the output, given in units

units
The unit in which width and height are specified

Value

The value of the return type depends on the type argument:

if type = "dendrogram", x is returned invisibly;

if type = "path", an object of class ggplot which can be plotted directly (by invoking its print method) or modified further by the user is returned;

if type = "dynamic_path", an object of class gganim is returned, and many be further manipulated by the user or plotted directly;

if type = "interactive", a shiny app which can be activated by invoking its print method.

saveviz.CARP always returns file.name invisibly.

Details

The saveviz.CARP function provides a unified interface for exporting CARP visualizations to files. For all plots, at most one of percent and k may be supplied.

Examples

if (FALSE) {
carp_fit <- CARP(presidential_speech)
plot(carp_fit, type='interactive')
}