Print CARP Results

Prints a brief descripton of a fitted CARP object.

# S3 method for CARP
print(x, ...)

Arguments

x
an object of class CARP as returned by CARP

...
Additional unused arguments

Details

Reports number of observations and variables of dataset, any preprocessing done by the CARP function, regularization weight information, the variant of CARP used, and the visualizations returned.

Examples

carp_fit <- CARP(presidential_speech[1:10,1:4])

#> Pre-computing weight-based edge sets

#> Computing CARP Path

#> Post-processing

print(carp_fit)

#> CARP Fit Summary
#> ====================
#> 
#> Algorithm:  
#> 
#> Available Visualizations:
#>  - Static Dendrogram:          TRUE 
#>  - Static Cluster Path:        TRUE 
#>  - Interactive Visualization:  TRUE 
#> 
#> Number of Observations:  10 
#> Number of Variables:     4 
#> 
#> Pre-processing options:
#>  - Columnwise centering:  TRUE 
#>  - Columnwise scaling:    FALSE 
#> 
#> RBF Kernel Weights:
#>  - phi =  0.1 
#>  - K   =  2 
#> 
#> Raw Data:
#>                     amount appropri  british     cent
#> Abraham Lincoln   3.433987 2.397895 1.791759 2.564949
#> Andrew Jackson    4.248495 4.663439 2.995732 1.945910
#> Andrew Johnson    4.025352 3.091042 2.833213 3.332205
#> Barack Obama      1.386294 0.000000 0.000000 1.386294
#> Benjamin Harrison 4.060443 4.174387 2.302585 4.304065