Print CARP Results

Prints a brief descripton of a fitted CARP object.

# S3 method for CBASS
print(x, ...)

Arguments

x
an object of class CARP as returned by CARP

...
Additional unused arguments

x
an object of class CBASS as returned by CBASS

...
Additional unused arguments

Details

Reports number of observations and variables of dataset, any preprocessing done by the CARP function, regularization weight information, the of CARP algorithm performed, and the visualizations returned.

Examples

carp.fit <- CARP(presidential_speech[1:10,1:4])
print(carp.fit)
Print \code{CBASS} Results

Prints a brief descripton of a fitted \code{CBASS} object.

Reports number of observations and variables of dataset, any preprocessing
done by the \code{\link{CBASS}} function, regularization weight information,
the variant of \code{CBASS} used, and the visualizations returned.
cbass_fit <- CBASS(X=presidential_speech[1:10,1:4])
print(cbass_fit)

#> Error: <text>:3:7: unexpected input
#> 2: print(carp.fit)
#> 3: Print \
#>          ^