Save CARP visualizations

This function will save dynamic CARP visualizations as gifs and static visualizations as image files

# S3 method for CARP
saveviz(x, file.name, plot.type = c("path", "dendrogram"),
  image.type = c("dynamic", "static"), static.image.type = c("png"),
  axis = c("PC1", "PC2"), dend.branch.width = 2,
  dend.labels.cex = 0.6, percent = NULL, k = NULL,
  percent.seq = seq(from = 0.05, to = 1, by = 0.05),
  dynamic.width = 1200, dynamic.height = 700, static.width = 8,
  static.height = 5, ...)

Arguments

x
a CARP object returned by CARP

file.name
name and location of image to be saved.

plot.type
a string specifying the type of plot to produce. 'dendrogram' produces CARP's cluster dendrogram and 'path' produces the cluster path

image.type
a string specifying the type of image output. 'dynamic' produced a gif of the plot.type along the CARP path. 'static' produces a single image of plot.type at a point along the CARP path.

static.image.type
string specifying the graphics device with which to save static images.

axis
a character vector of length two with elements as 'PC1','PC2',.. etc. Specifics which principal component axis to display for the 'path' plot.type

dend.branch.width
a positive number. Line width on dendrograms.

dend.labels.cex
a positive number. Label size on dendrograms.

percent
a number between 0 and 1. Specifies how far along the CARP path the static visualizations should display in terms of percent reguarlization.

k
an interger between 1 and n.obs. Specifies how far along the CARP path the static visualizations should display in term of number of clusters.

percent.seq
a vector of numbers between 0 and 1 specifying the positions along the CARP path used to generate dynamic images.

dynamic.width
dynamic output width in pixels

dynamic.height
dynamic output heigth in pixels

static.width
static output width in inches

static.height
static output height in inches

...
Unused additional generic arguements