Save CBASS visualizations

This function will save dynamic CBASS visualizations as gifs and static visualizations as image files

# S3 method for CBASS
saveviz(x, file.name, plot.type = c("heatmap",
  "obs.dendrogram", "var.dendrogram"), image.type = c("dynamic",
  "static"), static.image.type = c("png"), dend.branch.width = 2,
  dend.labels.cex = 0.6, heatrow.label.cex = 1.5,
  heatcol.label.cex = 1.5, percent = NULL, k.obs = NULL,
  k.var = NULL, percent.seq = seq(from = 0.05, to = 1, by = 0.05),
  dynamic.width = 1200, dynamic.height = 700, static.width = 8,
  static.height = 5, ...)

Arguments

x
a CBASS object returned by CBASS

file.name
name and location of image to be saved.

plot.type
a string specifying the type of plot to produce. 'obs.dendrogram' and 'var.dendrogram' produces the CBASS cluster dendrogram for the observations and variables respecitvely. 'heatmap' produces a heatmap of the CBASS solution.

image.type
a string specifying the type of image output. 'dynamic' produced a gif of the plot.type along the CARP path. 'static' produces a single image of plot.type at a point along the CARP path.

static.image.type
string specifying the graphics device with which to save static images.

dend.branch.width
a positive number. Line width on dendrograms.

dend.labels.cex
a positive number. Label size on dendrograms.

heatrow.label.cex
heatmap row label size

heatcol.label.cex
heatmap column label size

percent
a number between 0 and 1. Specifies how far along the CARP path the static visualizations should display via the amount of regularization.

k.obs
An interger between 1 and n.obs. Specifies how far along the CARP path the static visualizations should display via the number of unique observation clusters.

k.var
An interger between 1 and p.var. Specifies how far along the CARP path the static visualizations should display via the number of unique variable clusters.

percent.seq
a vector of numbers between 0 and 1 specifying the positions along the CARP path used to generate dynamic images.

dynamic.width
dynamic output width in pixels

dynamic.height
dynamic output heigth in pixels

static.width
static output width in inches

static.height
static output height in inches

...
Unused additional generic arguements