Compute CBASS solution path

CBASS returns a fast approximation to the Convex BiClustering solution path along with visualizations such as dendrograms and heatmaps. Visualizations may be static, interactive, or both.

CBASS(X, verbose = 1, interactive = TRUE, static = TRUE, control = NULL,
  ...)

Arguments

X
A n.obs by p.var matrix with rows the observations and columns the variables

verbose
either 0,1,or 2. Higher numbers indicate more verbosity while computing.

interactive
A logical. Should an interactive heatmap be returned?

static
A logical. Should observation and variable dendrograms be returned?

control
a list of CBASS control arguements; see cbass.control

...
additional arguments passed to cbass.control

Value

X the original data matrix

cbass.cluster.path.obs the CBASS observation solution path

cbass.cluster.path.var the CBASS variable solution path

cbass.dend.obs if static==TRUE, a dendrogam representation of the observation clustering solution path

cbass.dend.var if static==TRUE, a dendrogam representation of the variable clustering solution path

phi.obs A positive numner used for scaling in observation RBF kernel

phi.var A positive numner used for scaling in variable RBF kernel

k.obs an integer >= 1. The number of neighbors used to create sparse observation weights

k.var an integer >= 1. The number of neighbors used to create sparse variable weights

burn.in an integer. The number of initial iterations at a fixed value of (small) lambda_k

alg.type Which CARP algorithm to perform. Choices are 'cbassviz' and 'cbass'

interactive A logical. Should an interactive heatmap be returned?

static A logical. Should observation and variable dendrograms be returned?

obs.labels a vector of length n.obs containing observations (row) labels

var.labels a vector of length p.var containing variable (column) labels

X.center.global a logical. If TRUE, the global mean of X is removed.

Details

CBASS solves the Convex Biclustering problem via Algorithmic Regularization Paths. A seqeunce of biclustering solutions is returned along with several visualizations.

Examples

# NOT RUN {
library(clustRviz)
data("presidential_speech")
Xdat <- presidential_speech
CBASS(
X=Xdat
) -> cbass.fit
# }