Compute CBASS (Convex BiClustering) Solution Path

CBASS returns a fast approximation to the Convex BiClustering solution path along with visualizations such as dendrograms and heatmaps. Visualizations may be static, interactive, or both.

CBASS(X, verbose = 1L, interactive = TRUE, static = TRUE, ...,
  control = NULL)

Arguments

X
The data matrix (\(X \in R^{n \times p}\)): rows correspond to the observations (to be clustered) and columns to the variables (which will not be clustered).

verbose
Any of the values 0, 1, or 2. Higher values correspond to more verbose output while running.

interactive
A logical. Should interactive paths and dendrograms be returned?

static
A logical. Should static paths and dendrograms be returned?

...
Additional arguments used to control the behavior of CBASS; see cbass.control for details.

control
A list containing advanced parameters for the CBASS algorithm, typically created by cbass.control.

Value

An object of class CBASS containing the following elements (among others):

X: the original data matrix

n.obs: the number of observations (rows of X)

p.var: the number of variables (columns of X)

alg.type: the CBASS variant used

X.center.global: a logical indicating whether X was globally centered before clustering

X.scale: a logical indicating whether X was scaled column-wise before centering

burn.in: an integer indicating the number of "burn-in" iterations performed

k.obs: the number of neighbors used to create sparse clustering weights for the observations

k.var: the number of neighbors used to create sparse clustering weights for the variables

phi.obs: the scale factor of the RBF kernel used to calculate clustering weights for the observations

phi.var: the scale factor of the RBF kernel used to calculate clustering weights for the variables

carp.dend.obs: If static=TRUE, an dendrogram (object of class hclust) containing the clustering solution path for the observations

carp.dend.var: If static=TRUE, an dendrogram (object of class hclust) containing the clustering solution path for the variables

cbass.cluster.path.obs: The CBASS solution path for the observations

cbass.cluster.path.var: The CBASS solution path for the variables

static: a logical indicating whether static visualizations are available for this CBASS object

interactive: a logical indicating whether interactive visualizations are available for this CBASS object

obs.labels: a character vector of length n.obs containing observation (row) labels

var.labels: a character vector of length p.var containing variable (column) labels

obs.labels a vector of length n.obs containing observations (row) labels

var.labels a vector of length p.var containing variable (column) labels

Details

CBASS solves the Convex Biclustering problem via Algorithmic Regularization Paths. A seqeunce of biclustering solutions is returned along with several visualizations.

Examples

# NOT RUN {
cbass_fit <- CBASS(presidential_speech)
print(cbass_fit)
plot(cbass_fit)
# }